[roll] Roll infra/recipes [toolchain] Take builders as a property
We're preparing to rename all "fuchsia-*" builders to "core-*". To make
this as close to atomic as possible, it will be easier if builder names
are only specified in integration.git so we can do the switch in a
single CL.

Builder names generally shouldn't be encoded in recipes anyway because
it makes renamings like this difficult and makes the builder dependency
graph harder to understand, since it's spread across multiple repos.

Original-Bug: 75876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/523180
Original-Revision: 4eedf2b458c9d9b3f1e09e47ef95906e52011b98
GitOrigin-RevId: 0e7df43e15e93d84a5335b1aeda1bedc59c92e60
Change-Id: I92419679f21bb2c40964fa276e80c079235691df