[roll] Roll fuchsia [gidl] Add test case for vector of strict enums
This was failing in my in-progress CL and is an important case to cover.
In the C walker, primitive vector elements are no-ops while enums
validate the contents of the element. This means that a vector of enum
covers cases where the value is not a valid member that are not caught
elsewhere.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523100
Original-Revision: 3126abe3d4c5c1c0430624624935691c7a3263c5
GitOrigin-RevId: 26222707dc674ba78ff3bfafec05258b7bc270dc
Change-Id: Idc252ecd0e2a8821c6a93fda89023769eb99d593