[roll] Roll fuchsia [setui] Cancel and wait sensor task
Rather than canceling the sensor task directly, this CL adds a oneshot
to trigger cancelation and then waits for the task to finish completing.
This should help avoid a race where we drop in the middle of waiting
for the sensor to respond, which breaks the connection to the sensor
proxy. Also added an error log in the teardown state in case we discover
that the task is still running on teardown, which indicates other
problems with lifecycle management.

Original-Bug: 74576
Tests: fx test -o setui_service_tests
MULTIPLY: knowledge_card_test: 50
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522520
Original-Revision: d65cddd7270529f8a36dac7d8a3768c4ed666d86
GitOrigin-RevId: 67316e132d2369c8ddca88112303739ea8263b71
Change-Id: Ibc72e82654cfe9924d03dc8a2c5f333fbc23a539