[roll] Roll fuchsia [api] Clarifying 'default' value on tables
Using the term 'default' when referring to table fields is ambiguous as
it is unclear who of the sender or receiver is responsible for setting
the default. Instead, clarifying that the receiver is the one doing the
interpretation and considering absent fields as some value.

Original-Bug: 7889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523123
Original-Revision: c55e193cf2f2e2f0c4d64ee6cd6355520a144a64
GitOrigin-RevId: 5505f56078b65fabeeca61636e50b3cf896317ad
Change-Id: I2bf1305382a513fc9cbf52c987913f68f7035007