[roll] Roll fuchsia [mdns] fix reader bug - hang on name compression loop
This CL fixes a bug in which DNS message parsing can go into an
infinite loop due to a loop in name compression. That is, if a name
compression reference points to the reference itself, the reader
will spin forever. The root cause was a bad comparison (> rather
than >=) in the name reader.

Original-Bug: 62458

Test: Added unit test that fails before fix
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523300
Original-Revision: 4ea9cfdb0a69f677e28a099eb3b158fced5afe4d
GitOrigin-RevId: 63d5129059be6a316f430fd05c9f208a7e78d1c5
Change-Id: I3f4c5683338b88fe01b9c35770ab2b2821ad8ff5