[roll] Roll fuchsia [FDF] Exit the process using syscall.
While using the isolated devmgr v2 component in tests, intermittently
driver manager exits earlier than the driver host processes and the
driver hosts receive PEER_CLOSED events for each of the device channels
and for the channel between the driver manager and the driver host.

In this case, when we exit the process, we are ending up leaking the
memory of devices. Ideally the driver manager should gracefully shutdown
all the devices and then the driver hosts, and then kill itself. To
achieve this flow, the component lifecycle stop event handler inside
driver manager should
call into unbind hooks of the drivers and then remove the devices.
However this is blocked by 1.Untested unbind hooks in drivers.
2. unbind order still being parent first and child next.

This changeset temporarily calls the syscall for exit directly while we
get to the proper solution.

Other solution tried was to 1. Remove the device when channel connection
between driver manager and device closes. 2. Free up the dead devices
before exiting when the channel between driver manager and driver host
closes. However this causes failure in bunch of tests, that probably are
relying on the old behavior.

Original-Bug:75846, 75886

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522773
Original-Revision: 7f3e1e07a54e906dd9e9071848784fcdaafe80ea
GitOrigin-RevId: cbc8e7bd9c48bb45e231b405e66d19e47d216bc2
Change-Id: I251f8ce3d37a75d49f731038cd8c5ec2fc84840c