[roll] Roll fuchsia [blobfs] Fix tests in the new pager
Fixes CompressorTest and HealthCheckService tests using the experimental
Blobfs pager. The new pager is more sensitive to threads and exact setup
and shutdown conditions.

This uses the new test setup class to manage the setup and teardown for
the updated tests. THe health check tests also required that the blobs
be closed to avoid an assertion on teardown about leaking (normally the
VFS handles the open/close, and without this set properly the Blob
doesn't know to unregister itself from the pager watcher which caused
the leak. In the current (original) pager, there is another state where
the Blob can be registered with the pager but not kept alive by it.

The health check tests are the first test using the test setup that
requires blobfs be run on a background thread. A new object is set up
to handle this, and the existing setup functionality is moved to allow
it to be shared between threaded and non-threaded tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523289
Original-Revision: 0f7a42d48fa7bc008dd23e37fe68de174ff32fdc
GitOrigin-RevId: 5c45b15a11978a728e4472bb4d49de6feb6318b6
Change-Id: I4f1d82d59d5a81969e87e71240c319970e070f2d