[roll] Roll fuchsia [mtrace][perf_mon][x86] Add basic test of the mtrace syscall
Add a basic test of the mtrace syscall, specifically
MTRACE_PERFMON_GET_PROPERTIES. We expect perf_mon to be available
when the processor reports an architectural PMU with a supported
version.

There is other testing of the entire perf_mon subsystem, covering
the user driver, cpuperf providers, and so on; however there is no
focused testing of the mtrace syscall itself.

This test would have avoided a breakage in perf_mon fixed in
https://fuchsia-review.googlesource.com/c/fuchsia/+/513760 ;
it will also enable testing of the mtrace buffer lifecycle in the
future.

Tested:
$ fx set bringup.x64 --with-base //bundles/bringup:tests
  /boot/test/mtrace

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521816
Original-Revision: a59810a626bf34915d8f2da93d22063cf3c74ffd
GitOrigin-RevId: 7110ad97267079c3d90ca8f9511e6b084e339668
Change-Id: Ia433b1428cc2b82cb9f86cf59ee5c3a42b91972c