[roll] Roll fuchsia [virtualization] Update termina_integration_tests.cmx.
The current manifest fails to compile, and is missing two key
capabilities required for the guest instances to start.

Update the cmx file, and add the test package to the appropriate allow
lists.

After this change the termina tests are able to launch the guests, but
the tests still do not successfully start.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522901
Original-Revision: 5f03ea58478eeaefb61cc0e26cb3b9c93eb926a8
GitOrigin-RevId: 3b05b73238a008c805d7b1399016e8a07b71b983
Change-Id: Ie7a1ada1769b9ce9245f5aefb91973ef6a63b946