[roll] Roll fuchsia [pkdir-integration-tests] land framework
In this CL, we land the testing framework for pkgdir-integration-tests
(e.g. the pkg-harness and the rest of the test topology). In a
subsequent CL, we will create a unique test package for the framework to
use (right now, we're using the system image package).

Original-Bug: 75208
Test: fx test -o pkgdir-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522471
Original-Revision: e7e9a1b67730c3336717ce9c54ffcb3738d5f6b3
GitOrigin-RevId: 1d25d65b3a420363b36b4355f2e2e3c664f2eb29
Change-Id: I1195e52e0d455f98c96d21f617c6798d44bfab0d