[roll] Roll experiences [ermine] Remove deprecated use of ViewRefInstalled service.
Scenic use to require using this service to check if a view was
connected to the view tree, before requesting focus on it. This
restriction was removed recently. This change removes the usage of this
service from Ermine.

Original-Fixed: 61946
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/523500
Original-Revision: 560b8019cc3975c1ab2a62c7679b8e3499a6eb03
GitOrigin-RevId: 0fc7fa8289b1e0ed82b1d435e4236c9ff1c28924
Change-Id: I5b7155c3a56afc77247aafff40546982218e79eb