[roll] Roll fuchsia [cm][routing] Make check_use routing tests generic
This is the first of several changes making the
component_manager routing tests generic, with the
goal of testing both the CM router and the static
routing analyzer against the same set of test cases.

This CL moves a subset of the CM routing tests into
the routing_test_helpers library. The tests moved in
this CL are those that only need the check_use()
method on a test model.

Original-Bug: 61861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523153
Original-Revision: 8c2c3ecf42bb0203a2613b75ba4165efde1ab889
GitOrigin-RevId: 07dd9beb2db5e4144142a225ae36f5fffcd65abc
Change-Id: Ic943a18334542454c78d7afd7455140fa76c7567