[roll] Roll fuchsia [test-stats][devshell] Pipe timeout parameter to target.
The SSH process that is started along with the fx shell command does not
terminate when fx shell itself does. It seems that it remains connected
until the other side terminates or drops the connection.

In addition to possibly leaving dangling tests, this seems to break
tmux to the point that it is not handling input properly. Changing to
ensure that timeout is piped to the target device at least means that
the other side of the connection will be dropped when there is a timeout
(typically meaning the host side terminal is OK after execution since
the behavior seems to manifest when the other end is stuck on a very
long running test).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524228
Original-Revision: 0d493925bcf9bc04a15e1966de66c0d257ee7101
GitOrigin-RevId: db76a6dbab5f7d041f0a503b40548ec891191b76
Change-Id: I1202a10f38b32c7a2ec2388d1df795cc85b5f1c0