[roll] Roll fuchsia [artifactory] Don't list objs to refresh TTLs for if updated recently.
We have this check in the post-processor that updates the TTLs already,
but because there are a lot of duplicates, it needs to do a lot of
unnecessary Get calls without actually needing to update the CustomTimes
on the objects. Although we'd have to keep track of this in two places,
this should help speed up the post-processor.

Original-Bug: 66462
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523992
Original-Revision: 9664d483497e4f84f9082c68336b3232e0be15d0
GitOrigin-RevId: ccbfbfda7844fb532f970f8a3e94b3b573a06c15
Change-Id: I37ce1c0a3d9618aacec6bddb258df07c10a69702