[roll] Roll fuchsia [fxfs] Fix a SkipListLayer race
When searching for a node, there was a tiny window between the last
comparison and construction of the iterator where a node could have been
inserted which could result in returning a key that's less than the
search key.  The fix is simple: keep hold of the node that is used for
the last comparison and return that to construct the iterator.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523540
Original-Revision: 236d219641557e3749464ea377e06ae8084b2915
GitOrigin-RevId: b3345cd62a5a40d8de048d890741913a85a597d0
Change-Id: Iaae874b5d337242714d7d1e41a5007e8481c8c68