[roll] Roll fuchsia [audio_core] Move coefficient table creation out of filter.h
This is step 3 of a refactor to support fxrev.dev/471145.
In that CL, we need to reference coefficient table creation functions
from a source file that is not part of the Fuchsia build, meaning the
functions need to be defined in a header file that does not reference
any Fuchsia libraries except the C++ standard libraries. A follow-up
will ensure that mixer/coefficient_table.h does not reference Fuchisa
libraries.

I've also deleted the default ctors, which are unused.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512155
Original-Revision: 37837720aadea461c74349d1ed1dca99a36290a9
GitOrigin-RevId: 47e2509f902d4da333482ca60b89a395c3f339ff
Change-Id: Ide3cee05877834ca26bef51442f1afb2a95f815f