[roll] Roll fuchsia [build] Use a single toolchain "libprefix" definition.
The toolchain "libprefix" is used to determine the path of the
dynamic linker/loader, when generating Fuchsia executables, as well
as the installation path of shared libraries.

(For more details see the comment added in BUILDCONFIG.gn in the
patch).

This value only depends on the toolchain_variant fields, and was
re-computed in multiple places in our build rules. This CL simply
ensures it is only computed once in //build/config/BUILDCONFIG.gn
as toolchain_variant.libprefix, and converts all uses to refer to
this variable, with one exception (in sysroot_entries.gni below).

This simplifies our build rules a little, and provides a better
way to document this variable/value.

Apart from that, this should not change the build in any way,
and cherry-picking the CL should result in a no-op Ninja build
(after an automatic "gn gen" step though).

+ Remove toolchain.libprefix from
  //build/unification/global_variables.gni by replacing its
  uses with ${toolchain_variant.libprefix} instead.

+ Remove an obsolete comment in zircon_toolchain.gni

Original-Bug: 52720
Original-Bug: 61980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522300
Original-Revision: 9b164a51816d8cdf0e1ccb182f921cbacf8524e5
GitOrigin-RevId: 22d400cd05731f793033ead9d98cff672a08f368
Change-Id: I877c619d9146fd15683b266cdebba77c62c4e453