[roll] Roll fuchsia [pci] Use DDKTL for kpci
For some of the fragment work that needs to be done it will be
far easier if Kernel PCI and User PCI have a similar base. This
change brings forward the legacy C kpci driver over to DDKTL/C++,
but tries to make as few functional changes as possible otherwise.

Original-Bug: 32978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516720
Original-Revision: 1034678f320a978c9cf964993339f9c400231efd
GitOrigin-RevId: f82875b397d0d762d98cfa517a8777905dc630a3
Change-Id: Ie523f4846702a04575a49854582c09fd384ce72d