[roll] Roll fuchsia [fxfs] Fix a journal bug
There was a bug that prevented us replaying starting from the middle of
a block that has a reset checksum.  This CL changes things so that the
reset checksum works slightly differently: we pass a different starting
checksum to the writer rather than indicating reset separately.  This
does mean that detecting a reset now means we have to recompute a
checksum for an entire block, but that won't matter.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523780
Original-Revision: ea529080e7ab650d5088b5de9027ce519b07956c
GitOrigin-RevId: 0cdce4066648a05d812d58a79b21a9436e7733c1
Change-Id: I6faf0a14a8f5ada683d626e782e2b6ed1920323c