[roll] Roll fuchsia [pci] Implement PciProtocol in kpci
For the PCI migration there will be a period where drivers may need to
bind as composite drivers or as PciProtocolClients. To enable this,
Kernel PCI needs to support proxy devhosts as well as
PciProtocolClients.  This CL pulls the logic out of Rpc methods and
implements it within Pciprotocol methods, which are then called by the
Rpc methods. Rpc methods handle with any extra management
needed for the wire format for proxy communication.

There is no change to how the devices are added, or how they bind in
this CL.

Test: Manually verified via local builds and CQ testing, no automated
testing exists for legacy kPCI.

Original-Bug: 32978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517000
Original-Revision: 87aa2cab1ac85a42ef80250d77ec16a5c2ccf12f
GitOrigin-RevId: 4578230785ef8b24385073e30929b24ba8cc4fb6
Change-Id: I634c68737cc8d4af72f707a93e5c9d9fbaf0f3d9