[roll] Roll fuchsia [bt][lib] Correct generation bug in DetachableMap
In DetachableMap, you can get a DetachableWeak reference to an item, but
then that item could be replaced later.  If you then detached the original
item, it would remove the new item from the Map, even though you didn't
detach a reference to that item.

Add a bool tracking whether the item is in the parent map, and check it
was true before removing it.

Original-Bug: 75375
Test: added cant_detach_replaced_key
Test: fx test bluetooth-crate-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520371
Original-Revision: 668361c15c458fca181e91c32adadd3c14b32417
GitOrigin-RevId: db474d531faad28dd39e2715e5acc36c9eb76eb4
Change-Id: Ib2c78921b13a4586b0a3429408c5fc8649f84509