[roll] Roll fuchsia [bt][init] Remove IOCapability configuration from bt-init
The Control service now ignores requests to set the IOCapabilities
without a PairingDelegate set. bt-init was using the Control API to set
bt-host's IOCapabilities during stack initialization. There is no way
for a PairingDelegate to be set at this point, as the Control & Access
APIs for setting the PairingDelegate are not yet exposed to the rest of
the system. As such, this attempt to configure does not take effect.
Instead, we implemented non-bondable mode. Now, the system is in non-
bondable mode whenever a pairing delegate is not available.

Original-Bug: b/185940657
Test: `fx test bt-init-unittests` to verify this still builds.
      `fx test bt-profile-integration-tests` passed with logs which
      indicate that bt-init starts up successfully. I've also verified
      that these tests pass even with "stale" default.json config files
      which still have the "io-capability" field (i.e. that this field
      is ignored, not rejected, with the new code).
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521889
Original-Revision: a9b8381322b07bc5d26ce298e0045a6ea4beb1e7
GitOrigin-RevId: 97c4efd175952878f0607de9352b8757953c0b8f
Change-Id: I194fdc2cab4404db501955c48e808232821644dc