[roll] Roll infra/recipes [artifacts] Make upload step purple if it infra-fails
Slight refinement of fxrev.dev/524280 to make the "upload artifacts"
failure status (infra or non-infra failure) always correspond to the
overall status of the build. Otherwise the step would always be red when
it failed even if it turned the build purple, which is a confusing
state.

Original-Bug: 65555
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/524500
Original-Revision: f8e492155d55c8fe429f1794a05d8884275325e5
GitOrigin-RevId: e6d7c354a400adc55b4bba7a1964948c2bce9fce
Change-Id: Icebecf0568e1cd92ce4597ae252521a8076b06ed