[roll] Roll fuchsia [pci] Simplify zx_pci_bar_t
Remove the union so that we return a handle and address for all types of
bars. This reduces the friction in driver migration because this type
was already converted from zx_pci_bar_t to pci_bar_t to PciMsgBar during
the proxying.

Original-Bug: 32978, 74834
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523280
Original-Revision: 572a1442e1107baa2857ff1aa997ac170e17a564
GitOrigin-RevId: 19912c0f623465a6909cdc9454f88a511f207b3c
Change-Id: I6416473bc528cc31e1b79f7c5a896afbf48cfe4a