[roll] Roll infra/recipes [build] Emit failure summary as a Logdog log
Build failure summaries are often truncated in the summary markdown to
fit within the character limit, but it's important for users to be able
to view the full output without scrolling through hundreds of thousands
of lines of ninja output. The ninja recipe module does this by uploading
the full output to Logdog and linking to it from the ninja step, but I
forgot to implement an equivalent feature for fint.

Unfortunately we can't attach the log to the "fint" step because that
step will be finalized after we read the fint artifacts proto to obtain
the failure summary. So we attach the failure summary to the top-level
"build" presentation instead.

Also, trim trailing whitespace in the summary markdown so we don't
render unnecessary blank lines.

Original-Fixed: 75754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/525042
Original-Revision: db67c802854a8992625f126d36d93780a983631c
GitOrigin-RevId: ef4e2c814a4b86795656f0ac09ff810d3c1ea568
Change-Id: Ia8fec120c462b3ae6eaad35373f74172c5a6d946