[roll] Roll fuchsia [scenic] Improve DisplaySwapchain error detection.
The implementation has a comment that describes an assumption that
the swapchain will not be asked to recycle a frame-record before that
frame has been presented.  This CL makes it possible to assert the
truth of that assumption.

Logging is also improved.

Original-Bug: 73188

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521904
Original-Revision: d7b86515203c98f7c04da293cea2caf544cb2dce
GitOrigin-RevId: 543d77fb6c0d6b4820955799d3f003102b19784f
Change-Id: I9942a28902a4a5a9cbfb3d6f0f50239d06f2f8a6