[roll] Roll fuchsia [fidlc] Restructure type compilation
This CL adds a flat::TypeConstructorNew node that mirrors the
more general form that exists in the new syntax (i.e. a general
layout with parameters and constraints), and then introduces
a top level flat::TypeConstructor that contains either a
TypeConstructorNew or a TypeConstructorOld. All code is updated
to handle either kind of type constructor.

In order to properly support type aliases in the more general form,
compilation is changed so that Types are responsible for applying
constraints.

Original-Fixed: 75122
Original-Bug: 75112
Original-Bug: 74193
Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64-asan/exe.unstripped/fidl-compiler
Test: fx test --host fidlc_golden_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516460
Original-Revision: fdfcffccc9452064432c01e79b4de2776714be80
GitOrigin-RevId: 5f31463b9a34c46f74bfdeaa9563b2066008b754
Change-Id: I33b673f9ab3f3df22d9b6f4c1778dafb073695d6