[roll] Roll fuchsia [bindc] Add fuchsia.BIND_PCI_COMPONENT
To migrate the PCI Bus Driver from simple devices to composite devices
made up of multiple fragments we need a way to designate whether a
device driver has been migrated, as well as whether a device is for a
fragment or a simple device driver. This change allows us to mark a
driver as a simple or composite driver, as well as makes it possible to
denote a device as a fragment used only for building a composite out of
while other drivers still want to bind to it.

For this CL all PCI device drivers are "NOT_MIGRATED", aka simple.

Original-Bug: 74834
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519042
Original-Revision: ddad4fe8a4ab6ef2806c9a4bc1dc322063b51a98
GitOrigin-RevId: 4e164e81a38007e395da30ad701d7c28a4841b6c
Change-Id: I2d3771cf6be8cd5a9bbc1fc071124ab5efd1f827