[roll] Roll fuchsia [scrutiny] Add ffx scrutiny list package
This small change addresses a usability feature request with scrutiny
that makes it possible to list all the files in a given package from a
given package url. This makes it very easy to quickly inspect the
contents of a single package.

Original-Fixed: 74851
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525145
Original-Revision: 8f958aa867f2611ca6bac49d547f675c3a819651
GitOrigin-RevId: e2ed38dbd9d4fbc61951d2ed9bf39e82733948d1
Change-Id: I984e09ba0d631b23ef1e6d4fe93f29afb78cac36