[roll] Roll fuchsia [ffx][iquery] Allow Derivative on host
The iquery code uses the derivative crate to derive equality traits.
Currently iquery only runs on the target and therefore this constraint
was okay. However, I am in the process of generalizing the iquery code
to get some of it to also run on the host. This crate is used in some of
that host code so I want to make it available in that environment.

I couldn't find a compelling reason why it was restricted to fuchsia
only in the first place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525105
Original-Revision: 0d784126f0c2bac61851733c413bd411752d7d53
GitOrigin-RevId: f23f9e02fc8eb76203f2e937e983bd5111320982
Change-Id: I94d20a9b005a3ef46d19a2f1af23c560e247f96f