[roll] Roll fuchsia [fidlc] Refactor attributes to have flat representation
The purpose of this change is to refactor the internals of how
attributes are rendered in the flat AST.  The goal is to alter the
internal implementation only, in such a way that this new model can be
easily extended to support RFC-0086-style attributes.

Previously, attributes were one of the only raw AST node types to not
have a flat AST counterpart.  Instead, they were piped through directly
in their raw form, with no other changes. This was fine for the old
syntax: attributes could not reference any other declarations in the
FIDL file by name, as they were just a simple string:string mappings.

The new syntax allows for attributes to be constants, meaning that they
need a flat representation.  Attributes as a whole now have 3 flat AST
nodes:

* flat::AttributeList, which stores all of the attributes attached to
  a particular statement (`[Foo="bar",Baz]` in the old syntax, `@foo
  ("bar") @baz` in the new).
* flat::Attribute, a single attribute declaration (`Foo="bar"` in the
  old syntax, `@foo("bar")` in the new).
* flat::AttributeArg, the argument to an attribute (`"bar"` in the old
  syntax, `"bar"` in the new).  Note that for now Attributes may only
  contain 0 or 1 args, and that these args must always be
  StringLiterals, as that is all that is necessary to support all
  attribute declarations from the old syntax in the new one.
  Eventually, support for multiple args and all non-binary constant
  types (bool, numeric, identifier, etc) will be added.

Test: fidldev test fidlc
Original-Bug: 74955
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520023
Original-Revision: 49b951573e9272ef757068e63339c88cca7a1569
GitOrigin-RevId: a54d8bd56522d7b78132c09eb5b22d72ad64758c
Change-Id: I17c9b90c3c39dbb45193123450d38338edaaa90b