[roll] Roll fuchsia [volume_image]: Fix pave command line.
Length was supposed to be a default value when not specified,
instead of an overriding value. Renamed variable and switched the
behavior.

Also stat is not a reliable way to obtain number of blocks or size of a
device, so moved to last seekable offset.

Test: storage-volume-image-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525443
Original-Revision: 541e188d01a20c50cfd57da2f3f453d596dcec9d
GitOrigin-RevId: 7963f1505546b28919c40b1491df768a2247a0ed
Change-Id: Ifc9e172fe1e3b0a06c6b7440249b5c503ef829f1