[roll] Roll fuchsia [starnix] Flip FileCommon and FileObject
FileObject is renamed to FileImpl and FileCommon is renamed to
FileObject. FileObject is now the generic bits of an FD, and contains a
Box<&FileImpl> for the filesystem-specific bits.

The main reason I want to do this is to make the API exported by file
objects implemented by generic code rather than filesystem-specific
code. If the outer object is a filesystem-specific trait object, it's
harder to add generic code for a filesystem operation that applies to
every filesystem.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525448
Original-Revision: b31f5710481f64fde4e2655fe83aa53bda7ed143
GitOrigin-RevId: f452bbbadc5ee1c47fbe64f9a6163578e5af2130
Change-Id: Ifbf33281c97972957f3fda6e2498aa2a4345890c