[roll] Roll fuchsia [starnix] Create exception channel before starting process
Previously we started the process before creating the exception channel.
This was causing a race between the first syscall and the exception
channel being installed, which caused tests to flake when the first
syscall won the race.

Verified by introducing a sleep that caused tests to fail consistently
prior to fix and verified that they passed after the fix.

Original-Fixed: 75088, 75975
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526340
Original-Revision: 0b3392209f583eb7cfbdaba7fa0fc765d90191fe
GitOrigin-RevId: ed2383d626c0fc8a4d120f2e81ff62ceddaef2e0
Change-Id: I01feadca3b78eebea699230142ea338129442746