[roll] Roll fuchsia [memory] Generate crash report close to OOM
Write a crash report of type `fuchsia-imminent-oom` when the
Imminent-OOM kernel event is signaled. Unlike crash reports that are
generated at Critical, the Imminent-OOM reports are not rate limited -
OOM events are rarer than Critical crossings, and we would like to
gather information about each of them.

Note that not all Imminent-OOM reports might be followed by an actual
OOM, in case the system is somehow able to recover from memory pressure.
The Imminent-OOM reports are still valuable to have in this case, as
they point to atypical memory usage. If we find that a large number of
reports are being generated, we can consider rate limiting them in the
future.

Test: pressure_notifier tests

Original-Bug: 73872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513507
Original-Revision: 41c13a4f22a84b91fbcfe449d5e47c4cafedb877
GitOrigin-RevId: 9d931feeb87f30d5cb84bce74086b48c1594e6d4
Change-Id: I61ab5887a7b9b11c76cfc687f31c6ee730394449