[roll] Roll infra/recipes [no_op] Support custom summary_markdown
This will be useful to smooth out the process of renaming some builders.
We're leaving no-op builders with the old names in place for a few
months to prevent the builder pages from going away, and to make things
even more user-friendly we'll print a link to the renamed builder page
whenever one of the old builds launches.

Original-Bug: 75876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/525605
Original-Revision: 7c7bda674352555b12f56316fd59fab3b3d8268d
GitOrigin-RevId: 8d2e1f9e47e14ad47bc555b58dea4d8a7eafbaa2
Change-Id: I7fc90e120f9bd842260353f6510b96e006fc6861