[roll] Roll fuchsia [build] Add action for checking readonly vtables
As an alternative suggestion to
https://fuchsia-review.googlesource.com/c/infra/recipes/+/525703, we can
have a build action run at the end of every build that runs a script for
checking that all vtables are in rodata. This action is not used by default
but can be either run explicitly through ninja or used by adding
`check_vtables_in_rodata=true` to args.gn.

We would only like the script to be run on the clang canary
builders as a last line of defense from potentially breaking upstream
changes that could move vtables out of rodata. A followup infra change
will be added to add `check_vtables_in_rodata=true` to canaries as a gn
argument.

Original-Bug: 45314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526220
Original-Revision: ffefafe047702d8f1d4525ff2386ed288e3e9936
GitOrigin-RevId: 3d7aee6dfe6e6f3a60eb8a36cb2fdbc58bc76964
Change-Id: I98757f98e8e4eadf1fe709142ecf417ea170bd27