[roll] Roll fuchsia [kernel][arch] reset unsafe stack to the top before entering user space
On x86:
Previously, was not resetting the unsafe pointer when entering user
space for the first time, which was apparently leaking a bit less than
200 bytes of the stack which never get reclaimed. Subsequent entry/exits
from user space would never completely unwind the unsafe stack.

On arm64:
Not resetting the base of the shadow call stack on user space entry,
which could waste a few bytes of the shadow call stack. It's not exactly
the same situation as on x86 because the shadow call stack pointer (x18)
is restored to the saved copy in the arch portion of the thread
structure on kernel entry. This copy was only ever set on initial thread
creation and context switch, so it means that it'd only 'leak' the
memory if the thread was context switched at least once prior to exiting
the kernel for the first time.

TODO: consider on both arches just unilaterially reloading the
unsafe/shadow call stack to the base on every entry from user space.
Should be fairly easy to do on arm, but less clean on x86. May or may
not make things more ergonomic however.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523381
Original-Revision: c24948ad3544fe9ad7deeba82fb79139212f4f96
GitOrigin-RevId: d19b891dbff480bee82f21b9a58ca1e30eb653ac
Change-Id: I2cac5269d40eed48f773deca708cbd2677cb650a