[roll] Roll fuchsia [cm][routing] Make check_use_exposed_dir routing tests generic
Adds a method to the RoutingTestModel trait that checks a use
of a capability from an exposed directory and implements the
corresponding tests for both ComponentManager and for the
static routing analyzer.

Also adds a method to ComponentModelForAnalyzer to check a
capability use from an expose declaration when supported for
the capability type.

Original-Bug: 61861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524982
Original-Revision: a90f86be5e416a81e490c977daf18c6e288dee43
GitOrigin-RevId: 05e1ae8e900ce2b5d8bbbf3810d0d379d27fb383
Change-Id: I1d8a62094f5d0726fe171548c4ef4fe98e2b8884