[roll] Roll fuchsia [devices] Wait for drivers to stop when removing nodes
When removing nodes from the driver topology, if a node is bound to a
driver, wait for the driver to stop before removing the node.

This ensures that drivers can not be stopped out of order when run in
different driver hosts.

Original-Bug: 74571
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524684
Original-Revision: 2a78f221650ce10b3e1994c98fb17b05fbaa220a
GitOrigin-RevId: 06740d20e98125ae9ac2267369f429e76c7440ee
Change-Id: Idc227a0589e4292cb352a47a4a8e93a9b4af536f