[roll] Roll fuchsia [netstack] Remove hardError
The purpose of this structure is to leak errors out of loop{Read,Write};
it has hardly anything to do with being "hard". It is also incorrect;
these errors are expected to be consumed after being returned to the
user, which was not being done before this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525351
Original-Revision: 178d5e847b3c028af3b2b72ee8dfad4cfabedfce
GitOrigin-RevId: d0e65a63b24e835811c6041e38c08bf8319867c2
Change-Id: I6d81c56c63b169fb2ca89fcc622f9720a27539ab