[roll] Roll fuchsia [testing] Add proc macro to help with test fixtures
The #[fixture] proc macro takes one argument, which is the identifier of
a wrapper function.

The wrapper function takes two arguments: the name of the test function,
to be used or ignored however it wants, and a Fn of a signature matching
the input to the proc macro (the test body). The helper is responsible
for doing any necessary setup, and passing inputs to the test body
function. The macro essentially stuffs the test body into a closure,
which is passed to the test wrapper function along with the name of the
test.

For example, this test case:
```
async fn my_test() {
    let (proxy, service_fut) = setup_my_service("my_test");
    let ((), ()) = futures::future::join(service_fut, || async move {
        // test body using `proxy`
    }).await;
}
```

can be rewritten as:
```
async fn my_test(proxy: MyServiceProxy) {
    // test body using `proxy`
}
```

with this helper:
```
async fn setup_my_service<F, Fut>(test_name: &str, test: F)
where
    F: FnOnce(MyServiceProxy) -> impl futures::Future<Output = ()>,
{
    let (proxy, fut) = setup_my_service(test_name);
    let ((), ()) = futures::future::join(fut, test(proxy)).await;
}
```

Original-Bug: 75547
Test: fx test -d fixture-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521042
Original-Revision: 76a9a5f599d434c477606fce6942d9e3a2a12dff
GitOrigin-RevId: 7fc664c1faf0af1800ea23eb982925547e4deeda
Change-Id: Ic9a8a1f4d1d9fc8c101a1fe6795f50909c26012b