[roll] Roll fuchsia [sdktools] Updates to make SDK tools test friendly
This addresses two issues found when writing tests.

First, on some test environments, the SDK tools are
hermitcally added to the test by using symlinks to the
specific tool. As a result the normal path navigation within the
SDK does not work, and the real path of the executable should be
used to determine the location of other tools.

Second, added a new initializer to fconfig which allows for the
specification of the data directory vs. constructing it from the
user's HOME directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525144
Original-Revision: 1516b0e328badea38f030ce3ab000a064a300659
GitOrigin-RevId: bee405a3d4f102e760bc84490db774cfbe0e8d90
Change-Id: I0c6346f6566cf4a026d84c3d90d759fc0e3731e6