[roll] Roll fuchsia [blobfs] Add host support for reading multi-extent blobs.
This more closely unifies how the host and target read blocks for blobfs
but required some changing of the cache behaviour. The nodemap is now
held in memory as is done on the host so that it can use the same
libraries. This also meant changing the use of a dirty cache block for
flushing the inode. Instead the relevent block is just flushed directly
from the in-memory version, and since this was the only usage of the
dirty cache flag, it has been removed.

Tested: Manually.

Original-Bug: 48571, 32953
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524921
Original-Revision: 5998c79a065691df2ee44129573fca2796f6f166
GitOrigin-RevId: 44b9da3f4a00ae20e6954a456bd99f3176d28b7b
Change-Id: Ibc0cc1be5e656399f176f5c90aff4150b8c476e6