[roll] Roll fuchsia [debugger] Implement DW_FORM_implicit_const.
This is a new DWARF 5 tag. Currently it seems it is only used by GCC
which is why we had not encountered it earlier.

This was used to encode the size of an enum value, as well as for
DW_AT_decl_file attributes. This attribute is handled by LLVM at a
higher level than we use.

Handle this form in the DwarfDieDecoder by passing the form constant to
the form value.

The above change bypasses the setter for the DWARFUnit that we were
relying on in DWARFFormValue so the unit now needs to be passed around
in more places. This is probably for the better, though, because it
allows us to remove some const_cast calls and lambda captures.

Original-Bug: 76241
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526838
Original-Revision: 2755bd9dfef3d54e6e9d62dce2979438e575c0e2
GitOrigin-RevId: 658cfbf29d170fdf38806b571fbc153ff53d310b
Change-Id: I0498952abd4678c767bc48cc7972fb58277e91a2