[roll] Roll fuchsia [devices] Ignore peer closed errors in the DriverHost
When the DriverRunner closes the Driver channel, the DriverHost should
not report an error.

I was investigating whether there's a better way to do this, but there
doesn't seem to be a way to inform the server binding that this was
expected. In the other direction, server to client, you can send an
epitaph.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526548
Original-Revision: e29a7201a0fad70d7caa1d946ce1743b860b0f33
GitOrigin-RevId: 4b486f8576c3a819bc1b05cb8e6148b1bcc64705
Change-Id: Ib25bbbfc1b86631bea872ffcaeef1d353758123a