[roll] Roll fuchsia [multiboot] Don't leave booted (first) kernel item in data ZBI
The Multiboot shim takes a bootable ZBI and peels off the kernel
image from the first item.  Heretofore it then passed the entire
original ZBI as the data ZBI, so the kernel got a copy of its own
image as a data item.  This is confusing to carefully constructed
ZBI tests that expect to find other `KERNEL_*` items that really
are data items.  It's also not adhering to the ZBI boot protocol
as specified, which says that the boot loader delivers a data ZBI
including all the items *after* the initial `KERNEL_*` item, not
*including* it.  So change it to elide that item from the data
ZBI.  To keep it simple and not perturb the overall alignment of
the data ZBI, just edit the item header to replace the type with
ZBI_TYPE_DISCARD and leave it in place.

Original-Bug: 32255
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526697
Original-Revision: c1028e10c4ffb91a361bbad735459917e651ff7c
GitOrigin-RevId: bd676c4b413ffc5f9a63e5149c514381b9ed93b4
Change-Id: I4902be28ea99ce567b434f1103c201e20146241e