[roll] Roll fuchsia [kernel][ktl] Use a ktl-specific <__config_site> header
Use a local header to override the libc++ installation.  This
removes the need for a config() using defines for ktl.  It also
makes `#include <__config_site>` work again when the toolchain
installation changes so that its version is in a triple-based
subdirectory that isn't present for the modified triple produced
by -m32 for the kernel.phys32 environment.

Original-Bug: 75276
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527080
Original-Revision: 6d4d390da67f38e753a3fef07f7a8411622291f5
GitOrigin-RevId: 05c633041d9988e66415e68c8a0b76ca0696aa26
Change-Id: I83511da359959b2e5678e589cd251bc05c1c0442