[roll] Roll fuchsia [fxfs] Fix Admin/Shutdown
To avoid races in tests that immediately try and open the block device
after shutting down a filesystem, we need to ensure we've closed the
block device before sending the response to Admin/Shutdown.

In order to facilitate that, I needed to add a close method to
BlockClient, and in the process of doing that, I realised that it wasn't
properly async (the FIDL proxy we were using was synchronous), so I
fixed that and that involved separating out the synchronous and
asynchronous APIs.  Unfortunately, it does mean that Cache cannot be
used with a fake device, but that should be OK; we had nothing that was
relying on that feature.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525840
Original-Revision: 3098b56076d119f2172c033574a9ab3198139041
GitOrigin-RevId: 84609287451eac623c144468fcbb1142db96168f
Change-Id: Ie421bcac5ecdc040022c75042eaf26b5c60f36ea