[roll] Roll fuchsia [e2e-tests] Remove the update e2e test
Justification for removal: we have coverage of the same functionality
in other e2e tests, so this has become maintenance burden.

This update test in fact only tests that the 'update' binary functions
and talks to the update installer, which we have coverage of in our
other e2e OTA tests.

Original-Bug: 76034
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525044
Original-Revision: 93b24940a2982391c70cdfc082438a2811bd5010
GitOrigin-RevId: 20aff4b8b0efbea68b70434c70e61c9773df649d
Change-Id: Ibf4e2b3e5980369ba1f9dd2464c3026fb6f37826