[roll] Roll fuchsia [devices] Reduce the size of a Node
Reduce the size of the initial capacity of the FIDL allocator embedded
inside a Node so that the overall structure drops in size from 784 bytes
to 400 bytes. This means that 512 Nodes will take up roughly 200kb,
instead of 400kb. If the system allocator rounds up to the nearest power
of 2, the savings will be proportional.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526549
Original-Revision: 15cccef8a8b0587c45d654d2cf08f170d880e839
GitOrigin-RevId: ef927ecfa4dfba22c8b3928efa98af14de439b42
Change-Id: I49ddc0a6bd9af03a7b750401492557eac651b5c0