[roll] Roll fuchsia [bt][a2dp] Add v2 manifest
Add a CFv2 manifest for the A2DP component as well as an A2DP manifest
shard to be used by integration tests. This simplifies the smoke-test to
verify capability routing WAI and not component behavior. Defer the
declaration of the child component to the manifest so that tests can
hermetically include their own version of the child component.

Original-Bug: 73935
Test: fx test bt-a2dp-smoke-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521863
Original-Revision: 4acc13f4fa731aadb22480a2b489400593f10780
GitOrigin-RevId: 9527aba3844dee9c84e6bdae1460508a3dc73798
Change-Id: Ia04665325d7d8cb43ca60ac7ec6517bc3f1bfe97