[roll] Roll fuchsia [sysmem] Fix leak caused by GetAuxBuffers
Sysmem.HeapAmlogicSecure uses
fuchsia_sysmem_BufferCollectionGetAuxBuffers, which does
BufferCollection::CloneResultForSendingV2. The new collection info is
allocated in the TableSet, so even when it goes out of scope the VMOs
it references may still be held onto. This causes a reference cycle
to the LogicalBufferCollection, so it was leaked forever.

If we recreate the TableSet allocator in FailDownFrom we can ensure that
the VMO handles will be destroyed and the reference cycle will be
avoided.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526300
Original-Revision: 61f8565327785f747c7406d10c3bba2833941ea7
GitOrigin-RevId: b6f93bfcc55ac6deb2300b8674053c909631352d
Change-Id: I291c3f945b828f904fe2e3ec048e93aeaea0e4a4