[roll] Roll fuchsia [component_id_index] Fix monikers with nonzero instance IDs being rejected
The component ID index `moniker` field does not accept instance
identifiers. When the moniker string is parsed, it is converted to
an AbsoluteMoniker comprised of ChildMonikers that all have the
instance ID set to 0.

The check fails if the moniker of the incoming component has a
non-zero instance ID. e.g. with the index entry:

    {
        instance_id: "...",
        moniker: "/core/session-manager/session:session"
    }

component_manager returns the error:

    [component_manager] ERROR: bind() failed: Routing error:
    Component /core:0/session-manager:0/session:session:1 is trying
    to use a storage capability which is restricted to the
    component ID index.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525728
Original-Revision: 6b43057d2c1c884848b3dd4eced2ee0f7790b0f5
GitOrigin-RevId: be8a177f843a293164077f9a88a6930f752b6c7b
Change-Id: If45471e6ecee0e9385cf54f4cc53bb7160e1978d