[roll] Roll fuchsia [image assembly][build] Add support for json vbmeta descriptor files
Add support in the vbmeta signing template to take a list of json files
that describe the hash descriptors to add to the vbmeta image, instead
of vbmeta images that needs to have the descriptors extracted from.

This is a part of a soft transition.  The next step here will be to
remove support for the vbmeta hash descriptor extraction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526580
Original-Revision: 6421d6785ed1b4411a47dca0c97aea74c83c2492
GitOrigin-RevId: 1371e165aa2bfb79bda39f602c48b5d569c20600
Change-Id: Ibd126983a68bb243a301cde98cd6b521e15823f7