[roll] Roll fuchsia [fidlc] Refactor flat AST class hierarchy to add Attributable base class
All flat AST types that carry attributes have been refactored to inherit
from the new "Attributable" class as their base type.  This includes
Decls, [Bits|Enum|Service|Struct|Table|Union]Members, ResourceProperties
and Methods.

Constraint validation functions are now to const Attribute*, rather than
const Decl* as before (we also assert to ensure this pointer is never
null).  This allows attributes to be constrained in a more natural way
for non-Decl types (in particular, methods).  For example, in the old
setup, the way we verified the attributes for Decl::Kind::kProtocol
types was passing in their request/response structs as the const Decl*
to be validated.  This was unsatisfying, since what we actually wanted
to validated was the method, not just its arguments.

Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526080
Original-Revision: 20fe7bcc90cc18cb35969a88dc387fb673e1498e
GitOrigin-RevId: 8e587a7078687f82d06aedfd918f35a8494081d7
Change-Id: I7734f70c23a0da275dfebaffe51a7c136628c52f