[roll] Roll fuchsia [cmpctmalloc] Decouple max grow and alloc limits
This change rearranges some definitions with end result that the amount
the heap is able to grow by can more than the largest allocation that
it would actually satisfy.

In doing this various comments and intermediate definitions are added to
provide clarity due to internal metadata headers being included or not
included at various points in different calculations. This should
ensure that when the values for the maximum allocation and grow sizes
are changed in the future that incorrect/incompatible values are less
likely to be chosen.

Original-Bug: 30973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519890
Original-Revision: 2953916f1324539f6fa78a261caff71bb8a4d2a8
GitOrigin-RevId: dbccc036b581fd5e1be1a97d8add96fdebb79fca
Change-Id: I0af11b10087ed457ede03125c8b606fc9787fd09