[roll] Roll infra/recipes [gcc_toolchain] Don't hardcode builder names
To make it easier to determine the dependencies of each builder and make
renamings easier, builder names should only be specified in the
integration repo as much as possible. So pass in the builder names to
this recipe as a property (which is also consistent with how the
clang_toolchain and rust_toolchain recipes do it now).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/528183
Original-Revision: 559d7f2c78fbfb360a12659475d82438510c44d1
GitOrigin-RevId: 2941e8f75663da07ba24ce0c1ea82da55b7f2f9b
Change-Id: I29086c757ccc30049a659efa0f42b1ecb1185831