[roll] Roll fuchsia [run-test-suite] Add an integ test for the output directory
Note that the test is currently running on the target, but we intend to
collect output into a directory only from ffx on the host. Once we have
tests set up for ffx, we should move this test over to host.

Original-Bug: 75133

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525102
Original-Revision: e8665fe5cc2fb30595462662cc0716fcb593bb6e
GitOrigin-RevId: c0fa21ebdf59fc99559b2f000f717c9f92f201d4
Change-Id: I49c617fa659f267dafcd65d69d2a23a1e67f5c45