[roll] Roll fuchsia [debugger] Use a struct for process/thread koid.
Longer-term I want to make it easy to take a list of threads from the
debug agent that stopped and resume them. This was challenging with the
old setups, which for example had a process koid and then a vector of
thread koids in the resume request.

This adds a process/thread koid pair struct and uses it in the places in
the IPC protocol where appropriate. Also use it in some associated
non-IPC code.

Fix a problem I noticed in DebugAgent::OnResume where individual thread
IDs were ignored.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525882
Original-Revision: a0f2cce9dfcf5b2cedf45982c43f1add716dc8ad
GitOrigin-RevId: 527d5e0fbd6cac550a7fe66712a5f03a10c090e7
Change-Id: I21c4f28c8787f7670699a68c1872a94bad95bd98