[roll] Roll fuchsia [repro] Copy fx repro script into fuchsia.git
It was previously only available in internal checkouts because it used
to rely on `fx sync-to`, which is internal, but `fx sync-to` is now
optional when using `fx repro`, and nothing else in `fx repro` depends
on internal code.

The only changes I made to the script were to add the check to see
whether the `sync-to` command is available, as well as to add a `set
-euf -o pipefail` line and fixing a few broken pipes and unbound
variables.

See http://go/paste/6517091814342656 for the diff.

Original-Bug: 74724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528540
Original-Revision: 4af844d850fb4c2de7c061b7b540c9941b86d946
GitOrigin-RevId: 1a0c71d661327a153b7b39a2c7da9e5b4d692d1c
Change-Id: I9b1ee5a64d5468aff44e083ae2c87c6fb31387be