[roll] Roll fuchsia [assembly] Separate shareable vbmeta::sign
In order to also complete vbmeta signing in the `image` operation, we
are moving the shareable logic into its own file. This code does not
belong in src/lib/assembly, because it concerns argument parsing, which
is relevant to the ffx tooling.

Original-Bug: 76379
Test: fx test ffx_assembly_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527750
Original-Revision: 24e6f23c421755457f70135d0d54a0999150db98
GitOrigin-RevId: 1dbfcff8fd0d0837f451bf90c7c43abdbcdbc46c
Change-Id: I076a2f34ebb46ab2ee5bb085bf130040f5f00635