[roll] Roll fuchsia [ffx] fix timeout and port bugs in get-ssh-address
The timeout now allows for the target to be discovered after the command
is started. The output of the command always includes a port number,
defaulting to port 22 if none is set on the target.  Minor updates to
the error output string.

Original-Fixed: 75938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526944
Original-Revision: 4126e99eda1277f74561080e3753b2d8583d2f29
GitOrigin-RevId: 7ae537b64093bcc77ec6c4340fc3e36b4d448e19
Change-Id: Ie0b3073c648474f925eb40301edd7507e34b911a