[roll] Roll fuchsia [ime_service] Simplify keyboard focus events
Before:
The keyboard service tracks, for each subscriber, which keys had been
depressed when the subscriber lost focus. When it next regains focus,
the subscriber receives CANCEL events for keys it was tracking that
have since been released, and SYNC events only for newly pressed keys
that it wasn't tracking.
The subscriber is led to believe that the keys had remained pressed the
entire time that it was unfocused.

After:
When a subscriber loses focus, it immediately receives CANCEL events
for any currently depressed keys. When a subscriber gains focus, it
receives SYNC events for all currently depressed keys. There is no per-
subscriber tracking of keys. From the the subscriber's point of view,
no keys are ever in a depressed state while the subscriber is
unfocused.

TESTING:
1. Include `--with src/ui/bin/ime:tests` in fx set args.
2. Run `fx test keyboard3_integration_test`.

Original-Bug: 76086
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526507
Original-Revision: dad837859e0ba9aa92b8fe9f9bf5db7f7df30323
GitOrigin-RevId: 71cbd82510cc94392c9bda71961497b72e59fa32
Change-Id: Ib18330f2e900c0be3723e4bb70fb2c831fcb0c04