[roll] Roll fuchsia [fidl] Update DriverIndex.MatchDriver to support composites
This allows the driver index to provide the driver runner with
information about when a node matches a composite driver.

We use the `node_id` to specify which node was matched, and `num_nodes`
to specify the total number of nodes to be matched. This way, the driver
index does not need to hold any state, and the driver runner can make
the decision about when to create a composite node for a composite
driver.

Original-Bug: 73162
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528700
Original-Revision: ff26dfb54025984cb228551bbe8a38fc2acfc1f3
GitOrigin-RevId: 6558f89b7785f81f03a0c1e9698f19e85f2c091a
Change-Id: I88d5ff315d7b849fd8a1af45d5750b37b4b2677f