[roll] Roll fuchsia [fidlc] Refactor type shape calculation path for empty payload
Previously, a method request/response with an empty payload was "faked"
by being represented with an empty struct.  This allowed us to easily
calculate the type shape of the empty transaction header when necessary,
and was acceptable because the old syntax did not allow FIDL authors to
express that they wanted the payload to be an empty struct.  In the new
syntax, an empty struct is included in the domain of possible,
author-defined payloads, so this solution is no longer acceptable.

This change hard codes the correct type shape for an empty payload with
its transaction header included behind the TypeShape::ForEmptyPayload()
named constructor.  Call sites where we previously used an empty struct
as a synonym for "empty payload" have been altered to check if the
struct pointer is null, and call TypeShape::ForEmptyPayload() instead.

These changes caused some alterations in the goldens: the assignment of
`SomeLongAnonymousPrefix` names has changed, though in a way that
removed some previously erroneous cases that leaked into the IR.  This
is because certain cases which previously generated unused empty structs
are now short-circuited and return early (specifically: we no longer
optimistically generate an anonymous struct when there is no response
payload and no response error).

Additionally, the output of `declaration_order` field of some goldens
has changed, though not in a way that affects their topological sorting
or forward declaration order in C-family bindings.  This is because
methods that have empty payloads no longer point to "synthetic" structs
representing that payload.  This changes the shape of the dependency
graph between methods in a way that may affect iteration order, but does
not affect the final topological sort outcome (ie, all of the removed
nodes were leaves anyway).

Test: fidldev regen all && fidldev test fidlc
Original-Bug: 75526
Original-Bug: 68667
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526516
Original-Revision: cc748c2ad3f2d501a2ca5d9acacee29d094c7560
GitOrigin-RevId: ec3c6fe4fee232d00357dc94d2584002c2d8268a
Change-Id: Ic58df1bb13349014ffe5113daeb5751cd575e697