[roll] Roll fuchsia [blobfs] Remove old compression formats.
Removes read and write support for zstd, zstd-seekable, and lz4
compression formats from blobfs.

Old devices (blobfs minor versions <= 2) will no longer be capable of
being read with this change. Blobfs version 3 (now the oldest supported
version) is from early January 2021 and includes all versions of devices
in the field and all current branches. For the internal test devices
that have had previous versions of blobfs on them, they would have
needed to miss all updates for the past 5 months.

Metrics show 1-2 devices in this state in the past 35 days. But a device
that is reporting metrics but not updating indicates that it is broken
or stuck in some way already and will likely not be able to update to
this new version where the format becomes unsupported.

This does not report errors if the old bits are set. An original version
of this patch did that but that interferes with our requirement that
unused bits be not checked in case a future version uses them. I
considered keeping this check in for some period of time to catch old
devices, but I think we already have passed all the transition period
that is useful.

This keeps the FIDL compress enum definitions in them. I will look into
removing this in a followup.

Original-Bug: 69020

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528426
Original-Revision: c2645ee305844af3b3a736736f684e99f722140a
GitOrigin-RevId: 970c6f9971d7ea2c825447e30610a1ba56949b0a
Change-Id: I747c0283bc8ec3e3d46bfce84415412674f7dcbf