[roll] Roll fuchsia [starnix_test_runner] Keep component controller alive
This change keeps the component controller for the test component alive
until there are no more test suite requests for that component.

It also only serves one SuiteRequestStream, since it needs to close the
component controller on completion.

This is an attempt to fix a flake where a protocol is failed to be
routed, sinced the component "instance is not running."

Original-Bug: 76256
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528920
Original-Revision: a4b5617286e3cfdadf49788ab5bcdff5a30ac26f
GitOrigin-RevId: 9fcee4a8d1534ff7dd4bb406092ae68dc3f1f136
Change-Id: I76311ac416fd803c747e72b7871f49490069c0d7