[roll] Roll fuchsia [netdevice] Don't reclaim buffers on device stop
Modify device stop semantics to not automatically reclaim any in-flight
buffers. The existing automatic reclaim semantics are in fact not
enforceable. Replace that unenforceable guarantee with documentation
that implementations must return outstanding buffers when stop is
called, and that implementations must account for receiving buffers
while in the stopped state. The core driver still keeps track of
outstanding buffers, and will only teardown sessions (and thus the
device device instance) after all outsanding buffers are returned.

To properly automatically reclaim buffers -- and enforce that no buffers
are enqueued in the stopped state -- we'd need to hold rx/tx locks
across calls to enqueue buffers to the device, which we don't do to
allow re-entrant calls and also to avoid the assumption that these will
always be in-process synchronous calls, which may become an invalid
assumption when we support cross-process binding of netdevice.

Update serial-ppp and network-tun to conform to the new API requirements
that the implementation is responsible to return buffers when stop is
called.

Introduce test that exercises races between device stopping and rx/tx
buffers being pushed by the session.

Original-Bug: 76302
Original-Fixed: 76408
Multiply: network-device-lib-test: 30
Multiply: netstack_fidl_integration_test: 30
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526949
Original-Revision: 885e44a22db061fb550d6045d812c4400a12f368
GitOrigin-RevId: 22f75132bec861fcf33f5a48b1f395071d32dcbe
Change-Id: If8db06b5e9bfbb51b3ed00f0708c44946ede091e