[roll] Roll fuchsia [inotify] Make socket use datagrams and make mask type bits
This change makes the socket for reading/writing an inotify
event use ZX_SOCKET_DATAGRAM to avoid short reads/writes.
A read in inotify should consist of reading the inotify structure
in full or nothing at all. Additionally, a read should not discard
remaining inotify events buffered in the socket.

Similarly, currently we support an enum for mask of events.
However in linux, the api supports sending a combination of these
events together. Hence we need to make the type as "bits"
to allow this.

Testing: 1. Added AddMultipleFilters to test allowing a set of events
         being allowed by fidl instead of just one.
         2. Added DatagramPayloadNoShortReads to test for short reads/writes.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527680
Original-Revision: 901227aca77fff52da63eeed986f64092b332161
GitOrigin-RevId: 257e483431ad9d624833d1d9f0758e435a8f71ee
Change-Id: I961513de661c80ef59f8ee97834b303173031002