[roll] Roll fuchsia [ffx][pkg] Generate and serve repo configs.
Unfortunately the RepositoryConfig in fidl-fuchsia-pkg-ext isn't usable
on the host right now due to a dependency on fuchsia-zircon. I made
simplified types for this in repository.rs, which should be removed once
we can depend on the fidl-fuchsia-pkg-ext types.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524541
Original-Revision: 34e90d3821bad44359d6d9c6828e6e069f4f7536
GitOrigin-RevId: 48e8bd85933e54e09791ec0187e48dad08c92058
Change-Id: If70a750f291c1c2a804d115f666f2ffa45e89d93