[roll] Roll fuchsia [root-presenter] propagate VKManager notifications to VKController
When a VirtualKeyboardManager notifies the VirtualKeyboardCoordinator
of a change in the visibility of the keyboard, pass that information
on to the VirtualKeyboardController.

Also: add a FIDL test which verifies that a notification triggered by
a user interaction is received by a call to
fuchsia.input.virtualkeyboard.Controller.WatchVisibility().

Along the way, separate VirtualKeyboardController into an interface
and an implementation, to ease unit testing of VirtualKeyboardCoordinator.

Original-Bug: 73279
Original-Bug: 73281
Test: fx test -o root_presenter_unittests  -- --gtest_filter='VirtualKeyboard*'
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525035
Original-Revision: 40e8bd37979cb8942b6e6959ce987148976cf94c
GitOrigin-RevId: ba742078e96a2411f54d3e62d2d9b3494a73d083
Change-Id: Ief5257cc01e8a442b53b4df834fa0d988a4eac9c