[roll] Roll fuchsia [phys][page-table] Add address space builder for ARM64
Add a MapPage implementation for ARM64 and use this to create an ARM64
version of an AddressSpaceBuilder.

The builder will be used in upcoming CLs to create a 1:1 phys/virt page
table mapping, and longer term to map in the kernel at its final vaddr.

Currently the builder has a few limitations: it can only be used for
the TTBR0 address space (i.e., the low half of virtual memory), and
only supports a hard-coded set of attributes for mappings (in
particular, supervisor-only writability and the 0'th memory attribute).
These will be addressed in upcoming CLs.

Original-Bug: 67632
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513701
Original-Revision: a106244345ada48ae0d28f0ac6d8d6654a6369ee
GitOrigin-RevId: db48f7b06a62ed09b68547da3839df7273f99bbc
Change-Id: I3269c4f0bf09478399b443091d504f887b6af764