[roll] Roll fuchsia [fidlc] Disallow multiple constraints without braces
Previously this was not allowed, but we would throw an error
telling the user to surround multiple constraints in brackets.
Handling that error specifically is actually not possible since
it leads to a grammatical ambiguity.

This also removes the ErrUnnecessaryConstraintBrackets, which is
allowed according to RFC-0050

Original-Fixed: 76189
Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529041
Original-Revision: b6ee677522ceac0e9e5cec759d3cc8af60f40db4
GitOrigin-RevId: 9cbe5772b3c3130b9b3e7cdadf4b3ccdb55a4e5d
Change-Id: I859c73da9459c4af44ce0cac0b606e74b4a2b7d3