[roll] Roll fuchsia [ffx] get-ssh-address promptly sees targets
Manually added targets were not generating NewTargetEvents due to an
otherwised unused condition that does not fire events if the target is
unnamed.
get-ssh-address would take a minimum of 8s as a result of the poll
missing the event, but the timeout it used from get_target was 8s.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530047
Original-Revision: a1c770c36dd70c692d408bad7c080cca17223e58
GitOrigin-RevId: 70965c954d3cc05bdcfe3f38ad107063be8fc316
Change-Id: I9e5e531769e203f9d84d4f58f86c8c38fb7444da