[roll] Roll fuchsia [tools][fuzzing] Use tempfiles for Handle implementation in undercoat
In the previous implementation, the length of the serialized
representation of a Handle was directly proportional to the amount of
data stored in the Handle, resulting in some unwieldy tokens. This
changes the implementation to a more compact, tempfile-based one that
also gives us the benefits of 1) serializing type information so that
multiple types of connectors/launchers/etc can be used, and 2) being
able to store mutable data in a fixed handle if needed in the future.

Also updates a few paths to host tools and fixes a typo in a LSC
associated with the tools/qemu library.

Original-Bug: 47320
Original-Bug: 47479

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436748
Original-Revision: 737b7f2e1441b61c3f6e422e77798b08d0f01fea
GitOrigin-RevId: 90f8e713da826044391be3824be8912509a781d7
Change-Id: I97801650c02b9b955ed372f02ddc85d2c6bb3987