[roll] Roll fuchsia [archivist] Unified & Feedback reader is a v2 test
As a result a lot of duplicated code is now shared for both tests.

To be able to set /config/data in the test, a new configuration option
is introduced to specify the location of pipeline config. By default
this would be /config/data but for the test purposes it's set to
/pkg/config/data/pipelines.

Original-Bug: 76502
Tested: archivist-integration-tests-v2

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530462
Original-Revision: 8a0e2af6ecb01ccea6157e8a29e784ff74f2db13
GitOrigin-RevId: 18816d265b3e3267e9e0ea2cb4cdea5c1cfa2579
Change-Id: I807e72e1b75c235632434c31ac47e1058fad160a