[roll] Roll fuchsia [pkg-cache] Perform GC directly in pkg-cache
This moves garbage collection out of pkgfs and into pkg-cache. The
approach currently implemented is intended to be entirely compatible
with pkgfs. The only material difference is that, due to the NeededBlobs
protocol releasing blobs from the dynamic index after a failure,
incomplete blobs may be protected for longer in pkgfs than in pkg-cache.

Test: new tests added verifying parity between pkgfs and SpaceManager.
Original-Fixed: 66886
Original-Fixed: 66888

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514646
Original-Revision: 76ff6453463090669ee29af2fafc53340e2512e4
GitOrigin-RevId: 4b7ac6f1fb275127f16708287d2112b081bc3faf
Change-Id: I6513fd4095a44b79957992db6e6264999ff12fb9