[roll] Roll fuchsia [memory] Fix level bug in PressureNotifier
Make sure that the memory pressure level a Watcher of the memorypressure
service sees is one of Normal, Warning, Critical. This means that if the
current system pressure level is Imminent-OOM, it needs to be converted
to Critical for the Watcher.

This was a bug only on the callback and register paths. On the normal
level change path, we take care to not notify with the Imminent-OOM
level.

Test: pressure_notifier_unittest
Original-Fixed: 76980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529942
Original-Revision: 24d1a5d278713dfcd982700b5b0b317781e6b562
GitOrigin-RevId: 7fa169f75c2f28ca131b59997b2bfb1f19c752f9
Change-Id: I5a0d4cd9871ccb4242d42e022151e3c8e0fb6a81