[roll] Roll fuchsia [bt-init] Define temporary v2 bt-init component.
Define a manifest and separate binary for a v2 bt-init component. To
support the optional presence of bt-rfcomm, use CF APIs to attempt to
open the Profile service of the bt-rfcomm child component, falling back
to bt-gap's Profile service if opening bt-rfcomm fails.

The manifest's child URLs are absolute paths to production components,
which will not allow them to be packaged with bt-init for hermetic
integration testing. As such, this manifest will never be used as-is.
We will convert it to a CML shard which parameterizes its children (see
fxbug.dev/74547). It exists in this CL solely for making the review
process simpler.

Original-Bug: 72282
Test: `fx test bt-init-unittests-v2` to verify logic handling optional
      RFCOMM presence and maintain unit test coverage from v1. TODO -
      this is a step towards unblocking integration test coverage in
      CQ (fxbug.dev/53915), so this component will be thoroughly
      integration-tested before it is ever used in production. Also, we
      will write a manifest smoke test for this manifest once converted
      to a CML shard.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504726
Original-Revision: 7c73af8077017d2c56d23dc8bd53359714ff686c
GitOrigin-RevId: 15fb4c2149b04f54676b00773b560e90910c475d
Change-Id: I83ed7cf54b6a92a1dfea4d9fa74ad24a58ac0e7d