[roll] Roll fuchsia [ffx] add some CI target awareness to self-test
In CI, the leased target nodename is passed in via FUCHSIA_NODENAME, and
this is the target that we want to interact with. In other environments
where FUCHSIA_NODENAME is not set, attempt to discover a target in a
product state and return that nodename.

Original-Fixed: 76281
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530911
Original-Revision: 8e05a1dace694e5e5fcd451d00c51cb38be3f680
GitOrigin-RevId: eeb3b5bdb3cc1e51866608342902fe32c55fbd50
Change-Id: I737511232762b2bb2e13f6f0ada7792a08646cfd