[roll] Roll fuchsia [fidlc] Add support for box in new syntax
Adds support for box by:
* adding an intrinsic "box" entry to the new syntax typespace
* this entry points to a new template: `BoxTypeTemplate`
* this template produces new Type: `BoxType`

There are a number of approaches to then supporting the new
BoxType downstream: this CL does this by modifying the boxed type
(which is an IdentifierType that points to a Struct) to always
be nullable internally. This allows box in the new syntax to
follow the same code path as Struct? in the old syntax, which
is a nullable IdentifierType. This approach is more hacky as
it mutates an existing Type, but avoids the need to repeat the
same code twice. Instead of

```
switch (type->kind) {
  case IdentifierType:
    if (identifier_type->nullability == kNullable) {
      // handle nullable struct here
    }
  ...
  case BoxType:
    // handle nullabe struct here too by copy pasting code
    // from above
}
```

we end up with

```
if (type->kind == BoxType) {
  // type is now a nullable identifier type pointing to a struct
  type = static_cast<BoxType*>(type)->boxed_type;
}
switch (type->kind) {
  case IdentifierType:
    if (identifier_type->nullability == kNullable) {
      // both cases are handled here
    }
  ...
  case BoxType:
    assert(false && "should never get here");
}
```

Once only one code path remains (i.e. the old syntax no longer
supported), we can then consider moving the code for boxes
away from nullable identifier types of Structs (which would
then never be encountered) into the handler for BoxTypes
directly. In the meantime, this approach is chosen to minimize
the diff and the amount of copy/pasting.

Original-Fixed: 76192

Test: fx test fidlc_goldens
Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526761
Original-Revision: 6ff48f2060e950a599ebf8a4dc28b0c2c1d07b17
GitOrigin-RevId: a8c01e142a6b7a1ec8092239ba5b448601904bad
Change-Id: Id367910680b353406ed3ab82d0c409b1966e4895