[roll] Roll fuchsia [async-utils] Fix EventWaitResult drop panic.
Do not panic if the list of wakers does not contain the current waker.
The waker can be missing if the Event was dropped or signaled, waking
the waiters and removing them, and the Result was not repolled before
being dropped.

Test: Added unit test for this behavior
Original-Bug: 76728
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530511
Original-Revision: 83d6ea5c7899cec789be5f374f03e2fd6ea761ba
GitOrigin-RevId: 818d7903824fb662750f822402e2c084744c2ed8
Change-Id: I6545a83afefb58c809fbadcb0ef44e515f002dc8