[roll] Roll fuchsia [fidl][llcpp] Error logging support
We would like to hide `fidl::Reason` eventually from results and unbind
info. This implies that users shouldn't fetch and log the raw
`fidl::Reason` enum, which only results in a cryptic number anyways.

As a step towards that, this CL adds logging support to `fidl::Result`
and `fidl::UnbindInfo`:

- They can be logged to output streams, which includes FX_LOGS, gtest
assertions, `std::cout`/`std::cerr`, and so on.
- They can also be logged with `printf`-like formatters. However, this
is slightly less performant compared to the streams API, since we need
to build a temporary string on the heap first.

Based on discussions with the Diagnostics team, logging APIs are moving
to a structured-K/V pair model, but the structured API doesn't support
custom user types yet. When fxbug.dev/76211 is ready, we would be able
to make use of that. For now, string-based logging is the next best
thing.

An example of logging a result to syslog:

  fidl::WireResult bar = fidl::WireCall(foo_client_end).GetBar();
  if (!bar.ok()) {
    FX_LOGS(ERROR) << "GetBar failed: " << bar.error();
  }

When an error happens, the rendered log would look along the following lines:

  GetBar failed: FIDL operation failed due to decode error, status: ZX_ERR_INVALID_ARGS (-10), detail: non-nullable string was absent

Another example for when a `fidl::Client` unbinds:

  client/server endpoint was unbound due to peer closed, epitaph: ZX_ERR_INTERNAL (-1)

Original-Bug: 75702
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525500
Original-Revision: 3d19f8c50768cff3403ff98e084a6977d2d6209a
GitOrigin-RevId: 686f8488517bf8eeb5da104d4ead27ee9104c08f
Change-Id: I88ca394ed239ff8109bb40d7da1a271d87c722ce