[roll] Roll fuchsia [ffx][test] Create ffx test
Create 'ffx test' as a replacement for 'ffx component test'. For now,
this duplicates the logic in 'ffx component test' and adds it to
'ffx test run', with the exception of listing test cases, which is put
in 'ffx test list'. 'ffx test result' is added for new commands that
will allow a developer to inspect the results of previous runs and will
be filled in during subsequent commits.
'ffx test' is marked experimental for now.

Original-Bug: 75133

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528289
Original-Revision: 6ce35c317c97698760a2e99fe988859226b6f7af
GitOrigin-RevId: 5fc4118c9dd3e63873bf2ba7d0e82451fdb322a9
Change-Id: Ic48a3debd9dcba497ae15286d85411f27e526dc6