[roll] Roll infra/recipes [commit_queue] Handle Tricium analyzers
The recipe_testing module assumes that all the builders in
commit-queue.cfg are in the same bucket as the top-level builder.  This
has historically been the case, but is no longer true since I migrated
one of the Tricium builders into our main lucicfg codebase so it shows
up in commit-queue.cfg, causing recipes CQ failures such as
https://ci.chromium.org/b/8846932187768404496.

Tricium analyzers can be distinguished from other tryjobs by the
`mode_allowlist` field being set (generally containing "ANALYZER_RUN"),
so we use that to skip over Tricium analyzers.

Eventually it would be nice to be able to test Tricium builders in
recipes CQ, but we don't really need it now, and just skipping the
builders is the easiest way to fix recipes CQ.

(I also made a trivial change to the checkout recipe module so that
recipes CQ will determine that the tricium recipe is affected, to make
sure this fixes the issue.)

Original-Bug: 76644
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/531241
Original-Revision: 08786940192149fa4ae8cc96b66beeb89525d628
GitOrigin-RevId: ce2e70d4d473cb08c1c1479edb6880649e96e861
Change-Id: I94f66a2e17dd7263537bec6ecc329edc68adf5d0