[roll] Roll fuchsia [devices] Fix unsafe use of VFS in DirectoryFilter
Multiple threads can end up interacting with the DirectoryFilter, both
the main loop as well as the SystemInstance loop. This means that the
VFS that the DirectoryFilter hosts should be thread-safe.

This change switches over the DirectoryFilter VFS from a SynchronousVfs
to a ManagedVfs.

Original-Fixed: 76743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531203
Original-Revision: d6d2347819a6bd3ca3217f64a943d97efdac6e37
GitOrigin-RevId: 39b6bc6c9514480bdc5c7fa9a76e4ad7b8d30da6
Change-Id: Ie29524841df86a676fb8e76633d029dbe251cf3e