[roll] Roll fuchsia [pkg-resolver] Call NeededBlobs.Abort if a blob fetch fails
This guarantees that the pkg-cache removes the package from the
dynamic index before the resolve completes. This fixes a race
condition that could cause an immediate resolve retry to fail when the
pkg-cache's async removal of the package from the dynamic index occurs
after the pkg-cache has started caching the package for the second
time.

This also prevents the pkg-resolver from fetching any more blobs from
the affected package (it causes the NeededBlobs proxy associated with
the enqueued fetches in the BlobFetchQueue to immediately error).

The test verifies the canceled blob fetches, I manually verified the
flake fix by introducing a sleep in pkg-cache's serve_needed_blobs
to delay the async removal of the package from the dynamic index. This
sleep reproduced the flake 100% of the time without this fix.

Multiply: fuchsia-pkg://fuchsia.com/pkg-resolver-integration-tests#meta/resolve-recovers-from-http-errors.cm
Original-Fixed: 76372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530466
Original-Revision: 56d6c60d13da9cb52e36dee8ec53795bec611e48
GitOrigin-RevId: a8c8db4563cc60599ebae86d848f1d5935ffe58d
Change-Id: I713947d634c52bcc6bc86eddc7a27411c34de22b