[roll] Roll fuchsia [fidlconv] Correctly determine resourceness of request/response types
Previously, we'd only consider a top level request/response type a
resource if any of its members were directly a handle, but what
really matters is whether any member is a resource, which is a
broader requirement.

To do this, this CL augments the UnderlyingType class to use
it to determine resourceness whenever possible, but falling
back to looking at the flat::Decl when necessary.

An alternative approach would be to somehow store this data
in Library during compilation in a way that would be possible
to look it up during conversion (for example, mapping some
key of a raw::ParameterList to its eventual anonymous type,
allowing us to lookup its resourceness instead of re-computing
it based on the parameters). This would have the benefit
of avoiding to recomputing the resourceness at the request/response
level. However, given that fidlconv did not take its approach
for its existing resolution code, this CL tries to be consistent
with that which results in less code. Given the temporary
nature of fidlconv and strong test coverage, this approach
is considered to be good enough.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530868
Original-Revision: fa324cfb0e06a7249e95e0a1a113ab65170cc371
GitOrigin-RevId: 72ead89239c318b194958e9b78a6abd50f2898ce
Change-Id: I7116280da652f1f4c7ee8f0373dd002c42d7ffff