[roll] Roll fuchsia [ffx][keyboard] Adds the command `ffx keyboard`
This is an experimental command that allows us to configure the keymap
to use in the keyboard's input pipeline.  It only exists until the user
interface for the keyboard settings is implemented and  ready to go.

Removed the original `text_input_cli` since it is not possible to route
a request from a command line utility into a session without special
plumbing.  Instead, reusing all of that code in the context of an
experimental `ffx` command, as the session is accessible through the
`ffx` tool.

Original-Bug: 76057, 76058, 76056
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530509
Original-Revision: ed2a8cca42691f2d75be6577c8dcbcbcd606c64e
GitOrigin-RevId: b66554d3358d4edb0d67feb8a990a6b9511ca872
Change-Id: Id998fdbd7f32a2ffbc49b197fe4afc793d4bff30