[roll] Roll fuchsia [assembly] Allow clients to specify the bootloader name
The bootloader name is not always firmaware_<type>. In the current
build, in fact, it is almost never this format. To make comparison
between the current build and the image assembly outputs easier, we are
going to allow the name to be specified in the config.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531754
Original-Revision: a2abe1e2c847f5171ea622f4635cf5ccbce72a72
GitOrigin-RevId: d3285e62951b82800ed4f02aa5dc32427ea881aa
Change-Id: Ic161e4d3f67185daa8c46752b8e3690318dab494