[roll] Roll fuchsia [blobfs] Reland host support for reading multi-extent blobs.
Reverted in commit e99b65248d73d7e7dec5e89883e19abecfa0b299

The missing piece was lack of support for zero length merkle trees.

This more closely unifies how the host and target read blocks for blobfs
but required some changing of the cache behaviour. The nodemap is now
held in memory as is done on the host so that it can use the same
libraries. This also meant changing the use of a dirty cache block for
flushing the inode. Instead the relevent block is just flushed directly
from the in-memory version, and since this was the only usage of the
dirty cache flag, it has been removed.

Tested: Manually.

Original-Bug: 48571, 32953, 76514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531925
Original-Revision: 3c2c6aa71e359ce306602fe13627c05be29dde79
GitOrigin-RevId: 40221e7faf9b959f2c62a918f819f1e8e919dae6
Change-Id: I5c61f752377cdc7bd1fd66bdfee152beac754ebd