[roll] Roll fuchsia [ffx][iquery] Make iquery code a rust library
This introduces a rust_library to the iquery codebase which the binary
then depends on. As part of this change I tried to get as much as
possible available on host as well as target. There are still some open
questions mostly around files where things are only running on target. I
introduced a test which then makes it possible to ensure that the host
side stuff compiles correctly.

There should be zero functional difference from this change so the
primary testing story here is that the existing tests continue to pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525106
Original-Revision: 3d1631b37518402b656d0a603cf1e726a345c5d0
GitOrigin-RevId: 35c2b8b8d20b025cda60fab8816c0603667c4cf6
Change-Id: Ifabb714f1e8a14bdc2107117089df7c4a493b137