[roll] Roll fuchsia [system-updater] Migrate system-updater to a v2 component.
This change migrates the system-updater component from the software
delivery stack into the core realm as part of the v1-v2 migration
project. System-updater is an eagerly started component that provides
access to the fuchsia.update.installer.Installer service.

As part of this migration the component depends on several services
provided by other components in the SWD stack. These are routed through
core.cml to the component. In addition to this this component relies on
a set of features & directories:
- build-info
- deprecated-misc-storage
- isolated-persistence-storage
- pkgfs/system

Original-Bug: 73992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515223
Original-Revision: f9d4765bdfba30c99cc21fdaab38e1750df211fe
GitOrigin-RevId: d6f015858867c43e9461fac70ec72392568dc5cf
Change-Id: I8970c4207d1723b333ba1f6815211f7fe9bba509