[roll] Roll fuchsia [assembly] Generate salt file during GN-build
In order to compare the image assembler output with the GN-build output,
we need to use the same VBMeta salt. We are spitting out the GN-build's
salt to a file, so that it can be consumed by the image assembler to
produce consistent outputs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532381
Original-Revision: af51de9393ac28095a17e1bc4996eb51aeb9ae57
GitOrigin-RevId: 849047890166c31beed97dd62d5249bcb9cbcefd
Change-Id: I4db0330b04a9f9ed4e12ab6051ca1d60de3c87a8