[roll] Roll fuchsia [assembly] Disable passing boot args to ZBI
The boot args can be supplied via the IMAGE_ARGS field in the ZBI, or in
the devmgr_config.txt file. In order to enable binary diffing of the
GN-build ZBI, we are going to use the devmgr_config.txt for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532104
Original-Revision: 63f1e938666f955e6342e10a6d7e1b79130f8e91
GitOrigin-RevId: 059a9f4ac48bb0cc35e4508ddc49d5e0c29aca68
Change-Id: I9f48155599abb8be48d2e4769a0e9b3413b20ca7