[roll] Roll fuchsia [assembly] Remove whitespace from update_package.json
Removes the whitespace from the update packages' meta/package so that it
can be compared to the update package generated by the image assembler.
We are stripping the extension, so that `fx format-code` does not reject
the change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532075
Original-Revision: 0e996837b8f0b8f0091b4a55eb32cdf25e5c338b
GitOrigin-RevId: 4358605a4ffdac215e2def3fd0d37cdee4e04b05
Change-Id: I53ef1ea47ae204c8d26d1d905611504c41361839