[roll] Roll fuchsia [goldfish-display] Remove references to layer images if layer is unused.
If we unset a layer in SetDisplayLayers() for a goldfish
display device, we should remove all references of images
used in that layer in the driver so that the layer change
in display controller could be safely applied on next Vsync
signal.

This fixes an issue on session restart where Scenic creates
and sets a new layer for the display.

Original-Bug: 74677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532025
Original-Revision: 02c0dfff0fdba378f9df2093f3d4370c5fb509ba
GitOrigin-RevId: 1a401a33328066fc627530bdf772d28ef4d952b1
Change-Id: I16f34b5613395b91b4e1ef7a943e92eaad24cea7