[roll] Roll fuchsia [scenic][display] Remove DCHECK on unhandled vsync images.
For a certain DisplaySwapchain instance, it is possible that image
IDs in OnVsync() events doesn't match any of the images it imported
in the past; for example, it could be from other DisplayCompositor
using the same display controller FIDL connection. It is okay if we
simply ignore these images with "invalid" handles.

TEST: "fx shell session_control restart" on workstation.64 (no crash)
Original-Bug: 74677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532245
Original-Revision: a0568ab6d6395f07d8aa910377bff404ce584908
GitOrigin-RevId: 91689691355e7444c8f68f383a7d34f2f109e51a
Change-Id: I2a1a7ae8bf89ce1949ef67d1ca2b5a2bd5fe87e4