[roll] Roll fuchsia [boot-options]: Reduce stack size for tests.
Couple of times we've blown up the stack, due to the size of the boot
options and the multiple instances in each test. Alternatives such us
using the bootloader require some allocator for deletion which is not
available for physboot. On that note, we use static storage for a single
instance of boot options, and one accessor method, that resets its
content.

Test: k ut boot-options
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526161
Original-Revision: 004173ae083df260034dbd8b224d5aa89f9135d5
GitOrigin-RevId: 0114fa0b779b69bfc63a46dc1670d2fe0178219d
Change-Id: Ie8e54507ef919c57bbceb7877ba5ffd1aabf1954