[roll] Roll fuchsia [carnelian] Fixed BufferLayerCache cloning.
The cleanup logic for BufferLayerCache was incorrect since it was
upgrading the weak pointer to the SmallBitSet before the last
clone was dropped. This is fixed by using the other Rc in the
object to know when only the last clone is dropped.

This CL also contains a subtle fix in the clear logic: the
SmallBitSet should not be touched when clearing; it was introduced
by mistake in the last fix.

Last but not least, SmallBitSet is now not Copy anymore to reduce
the risk of accidental clones.

While testing, I've also removed a couple of debug prints added
previously by mistake.

Testing: fx test mold-tests & button-rs & layout-gallery-rs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532321
Original-Revision: 0712ce52a2a0598be739c40e781ba391422265a1
GitOrigin-RevId: 3eb0d86c10c71a7e9df890ab76dab0c26219f0ae
Change-Id: I99115d65a4ebcf1b5b4e4f3e487db85a20479e54