[roll] Roll fuchsia [zxio] Move zircon/system/ulib/zxio to sdk/lib/zxio
This moves the contents of the directory:
  zircon/system/ulib/zxio
to the directory:
  sdk/lib/zxio

This commit includes the following forwarding build target to ease the
transition to the new layout:

  //zircon/system/ulib/zxio:zxio

As this target is used by some rules outside the fuchsia git repository.

This commit is generated by the script //scripts/move_source/move_source.py and
should contain no behavioral changes outside a bug fix for the script
when trying to move files that have matches in checked-in binary files.

Several allowlists reference zxio and are also updated to refer to the
new location.

Original-Bug: 36063
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533021
Original-Revision: be9db25d7b348e201d1c7b2ab271a5c738ae7579
GitOrigin-RevId: ad21a5e8ef57842e0b9deb7873e8a4e9c5a298a3
Change-Id: Ib6ec654e3616b8a1a9942fa2c0b113074cd59b89