[roll] Roll fuchsia [blobfs] Change ownership of paged vnodes.
Change the ownership model for paged vnodes. This moves the owning
reference that the PagedVfs used to keep to the PagedVnode itself. This
way, the PagedVnode can always be registered for page requests for as
long as the VMO is alive without affecting its lifetime. This avoids
some race conditions identified in the added comments.

Use a separate lock for the Vnode tracking to prevent reentrant locking.
This was happening because Open in the Vfs creates the node while the
Vfs is locked, and then the created node is calling back into the Vfs to
register itself.

Fix the StreamingWritesTest by using the new BlobfsTestSetup.

Add a commented-out test for detaching vmos in the new pager.

All blobfs unit tests pass with the new pager with this patch.

Original-Bug: 76538
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530963
Original-Revision: c050ea6ca4c392ecfc4125cec1976540a99278d3
GitOrigin-RevId: e5ac631f430e67bea2ee635a1d9c081e9dc77da7
Change-Id: Ib63f84f34402c30bf565289a191c20265c19e2bf