[roll] Roll fuchsia [magma] Add ICD manifest json to package
In the future the loader service will read the manifest json from the
package and pass it to libvulkan.so in the client. This changes adds a new resource
target for that manifest to the icd_config template and updates the
Intel and Goldfish ICDs to use it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531681
Original-Revision: a42257a754c5770d46d81e0695a69d8ed26efbc6
GitOrigin-RevId: 1d2977e41665a51c47153bb131d9c1acde84586d
Change-Id: I440a4dfe98e8ed3b36c04d3e3c69fcf5204bf51a