[roll] Roll fuchsia [vulkan] Add feature query method to loader
The vulkan loader service in an image could be built at an earlier or
later time than libvulkan.so in a package. To help newer versions of
libvulkan.so deal with older versions of the loader service, add a
GetSupportedVersion method to query loader service features.

To help newer versions of the loader service deal with older,
potentially buggy versions of the client, add a ClientVersion enum to
the method. Currently the value will always be set to 0, but if bugs in
the client are found and the loader needs to work around them we can
increment the version number.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532162
Original-Revision: 6a830c157730ccf0d507db4ab006d098890e6900
GitOrigin-RevId: 411aefcf848e128241e2789506d3762607227f45
Change-Id: I80463f2a9193771bf83f4c5c392c76f2ae885e0b