[roll] Roll fuchsia [boot-options]: Kernel test is not thread hostile.
If kernel-test is compiled in kernel environment, it will use
ktl::unique_ptr, while if its compile in for phys environment it will be
static storage on the bss section.

Now this is hidden by a getter that provides pointer-like interface, so
tests only need to assume they have some sort of pointer. For static
storage, there is nothing to do, for ktl::unique_ptr it gets destroyed
when it goes out of scope.

To keep it simple, allocation failure, forces the getter to return
nullptr, and the caller asserts they got a non null ptr.

Original-Bug: 77209 # Done
Test: k ut boot-options
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533101
Original-Revision: f232c9d6026f09c57d4afda132b7d9acebb4a377
GitOrigin-RevId: 9bbf67e0c8d68f27248dbe63417003087f19a087
Change-Id: I1dd913612565e33d65a2bfa15e6ce3680431b9a0