[roll] Roll fuchsia [devices] Track nodes that have been orphaned
Nodes that have been orphaned during driver binding will now be tracked,
so we can provide a mechanism to implement policy.

In the future, we should provide a FIDL protocol in order to notify a
higher-level component of when a node becomes orphaned. Such a component
can then implement a device-specific policy, which may involve: hard
restart of the device, searching the internet for drivers, or requesting
user intervention.

Original-Bug: 76732
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532465
Original-Revision: 39e8a7090892956484d0d339daa03f561fcd3d81
GitOrigin-RevId: f0fbfebcb910e8d99942e2f79edfb3e96e01f1de
Change-Id: Ifb504715efe9ba2ad499938ccb037ec67b1424f9