[roll] Roll fuchsia [boot-options]: Fix unknown value handling.
Before this patch, running into an unknown value, would be a no-op
operation. This is an undesired behavior change from previous command
line library. The previous command line library would return the last
entry for a given option, no matter whether it was valid or not, since
the validation was delagated to the library user.

This patch introduces a default value for each member variable, and
updates Parse to return true when it successfully parses a value and
false when it doesnt.

Original-Bug: 53594
Test: boot-option-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526084
Original-Revision: 07c4a9da3ceac5b1e911689f8e3d68fae38f1d97
GitOrigin-RevId: 763f7168cef8a239154e745eefa373e9dc2da4b8
Change-Id: I83d21bf7d07da27606c335fbfcf3003f2a03e94e