[roll] Roll fuchsia [audio][lib] Update WavWriter for packed-24
The WavWriter library has been updated to accept SIGNED_24_IN_32
streams, but write packed-24 streams to file if the bits_per_sample
parameter to Initialize is set to 24. If it is set to 32, then the
library will write the 24-in-32 stream directly to disk as provided
to it by the client.

Bespoke code to support packed-24 has been removed from the utility
wav_recorder, as clients no longer need implement this themselves.

Parameterized unit tests are added, to validate file sizes and
sample sizes (both in-stream and in-file).

fx test -o audio-libwav-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530200
Original-Revision: a5432440faeb5273c03115e43577580b82f36095
GitOrigin-RevId: 6aef5d4375b330a13ca7859fadb89616b426264f
Change-Id: Iafd67306faf742cfd8c4f7248354eddcdf770033