[roll] Roll fuchsia [scripts][clang] Write script output to file
THe clang canary builders were failing because this script could run
multiple times without producing a no-op in the build. This updates the
script to accept a path fo a file and write to that file rather than
dumping to stdout.

This also updates the action that onvokes this script to write to an
file in the build directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532981
Original-Revision: aa68e334c3834c09b342f14100f76c044330a05e
GitOrigin-RevId: 397c1041fe3ccbd18fa180fe2eebae332913e9d3
Change-Id: Id231adce430e9b76187b541164033aeb15eec4f7