[roll] Roll fuchsia [debugger] Fall back to binary names when local.
When debugging on the local system, look for symbols in the binary
itself (based on the local path) if the build ID isn't found in the
normal index.

Currently this is not used but will enable local debugging in the
future. The detection is currently primitive (looks for "localhost") but
as we enable this mode local debugging will likely not require a
separate debug agent so it can be a direct in-process connection.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530146
Original-Revision: bef3443aff9f6ef3b033eb6f75a7f5efc1fd6af4
GitOrigin-RevId: f4d42be12bdae0d39faf04b7caaa901f139c2863
Change-Id: I0daa9eb49f74b07c66fda55737ead59e5ee3b459