[roll] Roll fuchsia [fuchsia-async] Make duration work on host
This change exposes Duration as a type from fuchsia-async the same way
that Time is exposed. This allows for platform agnostic timeouts to be
used.

The reason for this change is to let readdir_recursive work on host.
That is also included in this CL to help motivate this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532901
Original-Revision: 4ca03bba71c98690941b4a638639ea1fffdaa7ec
GitOrigin-RevId: 999f0c30aa668f91eb6e5b5c67f4b56f2bafb9fb
Change-Id: I23985136b5f9ef7e0c65530d5bac2e2e5372e185