[roll] Roll fuchsia [fx] Add subcommand "remote" to fx vdl and fvdl
As part of the new FEMU DX initiative, we're working towards unifying
how people use the emulator. This CL starts the process of adding a new
`remote` subcommand to the `fx vdl` and `fvdl` tools, which mirrors the
existing `fx emu-remote` functionality.

This CL introduces a placeholder subcommand, as well as the args that
will need to be supported by that subcommand. The subcommand is not
functional at this time, so we put a message in the help text indicating
that it's only a placeholder.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532582
Original-Revision: 2c2e8a30650ca575cb8ac15696b79f9194402e01
GitOrigin-RevId: a323c581487dd51fcd1cfebda2a400a799c067af
Change-Id: Id097cbf9561a7dab791ba86430099f6b036b83f5