[roll] Roll fuchsia [bt][cleanup] reformat comments
This change simply reformats comments to adjust to the line length of
the file. There are no other logic changes within this change. This
change is introduced mainly as a precursor to larger changes coming up
shortly. The change is intended to avoid noise in the larger changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533531
Original-Revision: 5a30ae6dfe33cf96bd09b2e34e7d40d2770391ff
GitOrigin-RevId: 1498eb81c3cd77121fc75f55a860eecabb8d6c3f
Change-Id: Ic2f0182a12576d805a9d8283545a819e309805a7