[roll] Roll fuchsia [debugger] Track calling convention on structures.
DWARF defines a calling convention for structures. This is used to
figure out how to pass and return values.

Read this from the symbol file ane track on the Collection symbol
object.

This information is displayed in sym-info but is not used for any
functional reason. It will be needed to display return values from
functions.

Original-Bug: 77333
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532764
Original-Revision: 5bf054134412cfaee7587235bff98a9eda3c0c78
GitOrigin-RevId: 6a24c76b8771f90cd18cfddc1b36254d34c3b60c
Change-Id: Iae2d2b87c8d22966fc2b5a038dad5cd464645f4e