[roll] Roll fuchsia [session] Add support for Atlas 4K display resolution
This change adds support for the Atlas 4k resolution.

It re-purposes an existing 4k 27" monitor resolution because Rust does
not allow defining additional monitor metrics with same resolution as
existing definitions. Since we need to support Atlas 4K officially we
are dropping the values for the 27" 4K arbitrary monitor.

This mechanism is anyways too crude to support arbitrary monitors. We
will need better solution to allow for that.

The magic value to scale the resolution on Atlas 4K was chosen such that
flutter reports physical display size of 3840x2160 and display pixel
ratio of 2.0

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533982
Original-Revision: d2ce4dc3459916166d079a0366142a79b87d9526
GitOrigin-RevId: e789a0149a7cce0cf5ba369f4781847eacb1598f
Change-Id: Ie36eff24ac246409529ec2667723375d78f9b172