[roll] Roll fuchsia [ffx] Convert target formatter tests to goldens
As part of https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=74157 I
am going to change some of the format of the output of target list.
These tests have a bunch of hardcoded formatted strings and a TODO to
replace those with goldens. This does that transition to use golden
files. The functionality should be the same so the goal here is that the
current tests continue to pass with this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533302
Original-Revision: 9a09c1917f7d376d5257b8871042a5d1e8c1a0ab
GitOrigin-RevId: 771e824bd376271a55e5922c53b9191eb83403b8
Change-Id: I50ed401f4d7ecd491efb39f6525534333c11399c