[roll] Roll fuchsia [covargs] Merge entries from all symbolizer output files.
Multiple files could contain entries with the same name (which is the
name of one of the profraw files) but with a different set of modules
and segments. These would get overwritten with whatever's in the last
file that references that profraw file which would cause coverage
collected from previous files to be missing if the same modules weren't
seen anywhere else.

Original-Bug: 77034
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534161
Original-Revision: a93f44389800d7f54e9d03f8470893cf3737b4fd
GitOrigin-RevId: 8046cc2280417bff6a1cd4721a7301f550ca652a
Change-Id: I499fd66bfbb71ff5677fa0ed8b6a18bf976044df