[roll] Roll fuchsia [soundplayer] Route LogSink
Follow up to fxr/504523.
Noticed this because I saw an error log about soundplayer using logsink
but not being offered it.

It's not apparent in fxr/504523 that soundplayer uses LogSink and
therefore is missing a route because the use of LogSink is slightly
obscured by the fact that it's done in the logsink manifest shard.
This specific issue is not covered in the migration guide.
I'm not sure if ffx scrutiny would have caught this or not. This may be
shadowed by fxbug.dev/77336 (speculating).

How can we do better in the future?

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533044
Original-Revision: 205a7004ddf91069ad919820963819b29d18aa6d
GitOrigin-RevId: efe13a16cfedec49d27a7e6016a31f7250116560
Change-Id: Ie76be5bd7b618ed5a7d610c07b2c1eaf05c5efc0