[roll] Roll fuchsia [ffx] Change target list to have multiline IPs
This introduces the concept of a field that can be either a string or an
vector of strings. If it is a vector of strings then each string is
printed on it's own line. This shrinks the size of that array column to
the maximum string length over the array.

https: //bugs.fuchsia.dev/p/fuchsia/issues/detail?id=74157
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533303
Original-Revision: 0dda6f2d1d683c56b7b70dc66289ac2c13b7c964
GitOrigin-RevId: e471c54cee2c85034a094fc09e8f59f6a1782464
Change-Id: Ib4f92b0c240e901aafe4ba2888f00ec019169ac2