[roll] Roll fuchsia [vfs] Use a separate lock for tracking vnodes.
As with the PagedVfs tracking of registered vnodes, the Vfs tracks open
nodes (this duplicate tracking can hopefully go away when the paging
capability is completely landed and merged with everything).

And just was with the PagedVfs, the Vfs needs a separate lock for
tracking opened nodes to avoid reentrant locking (nodes are destroyed
and created from within the main Vfs lock, so they can't call back into
the Vfs to register with the same lock).

This makes most of the integration tests pass and allows the system to
write new blobs at runtime (previously this would fail so no tests
could be run on a system with the new pager enabled).

Original-Bug: 51111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533262
Original-Revision: d359272ac9da25b13b6a1b628641134cbc1379a7
GitOrigin-RevId: 3e24b448d1d741805de66f46ce8270a014b69662
Change-Id: I0737f5e7bced23e405cf37beff51d0e23a6c64f3