[roll] Roll fuchsia [omaha-client] Fix flakes when subtracting instants
It's possible that `Instant::now()` is the number of monotonic seconds
after a device is booted. This means that subtracting a value from
`Instant::now()` can risk an underflow panic, especially if the test is
run on a device that was just booted.

This patch fixes a flake in omaha-client by making sure that whenever we
subtract a value, we first advance the `Instant` value by 24 hours
before subtracting from it to make sure we don't trip over this
underflow.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534522
Original-Revision: c1dbbd291e6ced94313daf26ad1f312826762469
GitOrigin-RevId: c4475bea2dfba7d566c2de052f62961ebfa6dae4
Change-Id: Ia54e0e4f64b23f0287ca03ba4e6a56213bdcddc9