[roll] Roll fuchsia [bt][profile-test-server] Use salt for service registration
Context: When a peer in the piconet decides to advertise a service, the
profile test server processes the service advertisement, extracts
relevant BR/EDR information from the service record, and assigns a
unique handle to the record for the lifetime of the record in the
server. This unique handle comprises of the peer's unique PeerId and a
unique number (from the Slab-backed ServiceSet object) assigned to the
advertisement. Together, it is guaranteed to be unique for the lifetime
of the record - this is called the `RegisteredServiceId` in code.
When the peer decides to stop advertising, the service record in
the profile test server is removed. Because these records are stored in
a Slab<T>, the handle used by it _can_ be recycled. In fact, based on
the Slab implementationn, it's very likely that the handle (which is
just a usize) will be reused.
If the _same_ peer decides to re-advertise the service, it is possible
(and probable) that the Slab's handle will be the same, resulting in the
same RegisteredServiceID being re-used. This means that any searches by other
peers in the piconet may _not_ be notified of this re-registered
service (because they filter by the RegisteredServiceID to prevent
duplication), even though they should be.
Adding a salt parameter to a registered advertisement allows us to
mitigate this issue.

Test: Added to bt-profile-test-server-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534232
Original-Revision: 497269cd655a08cce01d61257d671f895af5e84b
GitOrigin-RevId: 0fdfa530b58294aec1d964909f71afde42175810
Change-Id: Iccd0a9ce356f661c4b4bb0ff03b37036abc56227