[roll] Roll fuchsia [ftl] Reject map pages with all 0xff in the second half of OOB.
This rejects map pages with all 0xff in the second half of the OOB in
order to reject partial page writes which result in wear count listed as
the max value and the validity bits are 0xff, though validity bits are
currently all 0xff.

This also starts writing a form of checksum to the last validity byte to
ensure that the write completed, and to spot any unexpected 0xff
patterns in the oob that may have resulted from an incomplete write.

multiply: corrupt-tests (fuchsia): 20
Original-Bug: b/179400686
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529803
Original-Revision: f38989c02e281be8a362ba7412b85f60fbc46ef4
GitOrigin-RevId: e6a0a7b2daf3ca18c8b735ec70d8a42776a08ad6
Change-Id: I97fc159131f3977fb4bdcad3dc35957cf5f39613