[roll] Roll fuchsia [modular] Don't dispatch TestHarness events on launcher shutdown
TestHarnessFixture tests wait for the TestHarness to shut down
as part of TearDown. This may cause OnNewComponent events to be
dispatched to FakeComponents that were created in the test and
destroyed, or that are using variables in the test scope.

This prevents the FakeComponent OnCreate handler being unexpectedly
invoked after the test is finished. This may happen if the test
restarts Modular, for example.

Also, don't terminate TestHarness when a client disconnects.

Original-Fixed: 77109
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533649
Original-Revision: 0b2b956967850438a32dc7763ae5269cf5c7c8b4
GitOrigin-RevId: de3a12f05fe739da916f29c8eb858b9e852f0810
Change-Id: Ic01599e76fdd355484140e38aa9fd017eb6ea5f7