[roll] Roll fuchsia [fxfs] Some initial changes to support allocator reservations
We need to be able to reserve space for the journal so that writing to
it should never fail with an out-of-space error condition (for which
we'd never be able to recover from), so we need a way to reserve space
for it.

To support this, I changed associated objects slightly so that they
support both owned an borrowed objects.

This includes changes to the journal such that it will pause new
transactions when waiting for a compaction to finish to relieve space in
the journal.

I've moved the handle out of JournalWriter which means that we no longer
need to hold the writer lock when trimming space at the beginning of the
journal.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533284
Original-Revision: 0975f53e4eb55c4e898ab456bc30dce8a7e379b9
GitOrigin-RevId: 03da33c2dba6bce034e5ee534ee917cd4f86c555
Change-Id: I1af6e10a83be96c2cab0c396b02672bf23461af1