[roll] Roll fuchsia [audio][aml-g12-tdm] Use frame size to adjust mclk
The AMLogic g12 TDM clock configuration needs to change when we use
different frame sizes. This was found to be needed for the PCM output
used for BT HFP support. In BT HFP we use (depending on configuration)
a one channel 16 bits slot PCM configuration with a frame size of 16
bits. This in contrast with I2S that uses 64 bits frames.
In this CL we adjust the mclk configuration based on frame size to get
the correct clocking setting.

Original-Bug: 76759
Test: aml-g12-tdm-test
Test: Logic analyzer check of PCM streams at different frame sizes
(16 and 64 bits) and different frame rates (8, 16, 32, 48 and 96 kHz)
on an Astro "Oportus Edition" with PCM pins exposed.
Test: Manual audio check on Astro and Sherlock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532061
Original-Revision: 7a3c9d39de609be5969ecb361fa4d2acc2f26767
GitOrigin-RevId: 5a064190e78e690247f7e0940bb2b995f815517d
Change-Id: I89614bba5aa25d0299c4de1371e29cc1e87a7925