[roll] Roll fuchsia [audio][aml-g12-tdm] Allow multiple mclk pad usage
When using multiple audio engines as in TDM_B for speaker output
and TDM_A for BT output, we need to allow for different MCLKs to
be enabled simultaneously. With this CL we won't override the MCLK
configuration of one engine with the mclk configuration of another.

Original-Fixed: 75872
Test: aml-g12-tdm-test
Test: Manual audio check on Astro and Sherlock
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532064
Original-Revision: ec2918bd21cb679338bc9ca45b4e7289b103ade7
GitOrigin-RevId: ae1c57e6bdea26735baee8a9b62d852ac297998d
Change-Id: Ia3cf2656614c2636ff0e827bfb4c3eded134611e