[roll] Roll infra/recipes [increment_version] Add recipe to increment versions on branches
This recipe will be used to increment versions on release branches when
a no-op is needed to retry build(s), in cases of unexpected failures
that can't be fixed by a change to the source tree.

An example of when this is needed is when an infrastructure bug breaks
a build, and a new build is needed.

A plain retry with the same version has the bad side-effect of breaking
the guarantee that a version has no more than one corresponding build
for each builder, so we opt to create a new version here.

This replaces the hack that we've been using i.e. cherry-picking and
reverting whitespace changes.

Original-Bug: 77466
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/534941
Original-Revision: de92c68b38714b3002adc4fe7903d0727db5a077
GitOrigin-RevId: 3ef648d8718d1bd34e04d7e8860719814ed660c3
Change-Id: Iafba3de4dd1b84f593e0f48c8cf570631ddcdedb