[roll] Roll fuchsia [time] Update frequency documentation.
The original algorithm was a bit inconsistent with the handling of
frequency, occasionally working with utc per monotonic and occasionally
working with monotonic per realtime. The implementationation is
consistent in only using utc per monotonic and this CL cleans up the
documentation to match and be more explicit about its use of frequency.

I'm also removing the note that frequency isn't implemented yet to
celebrate the fact that we finished it!

Original-Bug: 56868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534981
Original-Revision: ba313a26ed83d98c42953b8f3ce8f40779c8048a
GitOrigin-RevId: c4c065078e635f46fad273c0506a18913b3fa347
Change-Id: I3a88f1fbb39aaf75bd41c847af50b70fda8604a1