[roll] Roll fuchsia [sdk-tools] fssh, fserve respect configured sshport
Updates fssh and fserve to check for a custom ssh port being
configured via `fconfig` if the --device-ip flag is not used.

The device-name is used as the key to read fconfig values, so
if --device-ip is given, there is no way to look it up.

The method RunSSHCommandWithPort is a transitional method to allow
migration of other tools that use this library in other repos. This
method will be renamed RunSSHCommand when the migration is completed.

Original-Fixed: 77410
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534885
Original-Revision: 8fb1879239aff9bd37f4953bd42419fa7069b110
GitOrigin-RevId: 439c550176de0be23fc0dec14d2ec885912d083e
Change-Id: I62d3673e4982075bafcd81ee9155b15c7cfb333c