[roll] Roll infra/recipes [affected_tests] Fix never-affected tests list
The "name" field in tests.json used to contain each test's base name
(e.g. "foo_tests") but now it contains either the test path or package
URL depending on whether the test is a host test or Fuchsia test. So
instead of checking the name, check the GN label, which should be fairly
stable.

This must have gone unnoticed for quite a while, so maybe that's a sign
that we no longer need this deny-list?

Original-Bug: 75371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/535601
Original-Revision: 698c35c4e2193e5c52e655da8dca2ceed3ac9fe1
GitOrigin-RevId: 366dbfaf691e4e01d4d719447823b419dd6e0a21
Change-Id: Ic3646d907253da5a6084786eae2a147a22fcbfa4