[roll] Roll infra/recipes [fint] Analyze affected tests
Logic translated from the affected_tests.py script in recipes:
https://fuchsia.googlesource.com/infra/recipes/+/1d3ba67d813b21c10a37e785ad66abf9a643e742/recipe_modules/build/resources/affected_tests.py

The ninja dry run output is now discarded after use and won't be
uploaded to Logdog, which will admittedly make it slightly harder to
understand the analysis. However, the dry run output is ~2 million lines
long for most builds, so it takes a long time to upload to Logdog and
isn't particularly useful anyway, because Chrome is extremely sluggish
to even scroll through the log, let alone search through it.

One thing that could be described as a regression is that fint will do
this analysis even when running in a build-only build (where we won't
run any tests) or an integration change (where we'll run all tests even
if the build graph analysis indicates no work). If this becomes an issue
we can add a context.proto field to let recipes control whether we do
the analysis. However, it only takes ~10 seconds currently, and since
all the slowest builders in CQ run tests (so they use the results of the
analysis), skipping it in other builders would have basically zero
impact on CQ durations.

Before updating the recipes to use the affected_tests and no_work
results emitted by fint, I plan to add a check that makes sure that the
results are the same as those emitted by affected_tests.py and only
switch over once the check has been running for a few days with no
failures.

Original-Bug: 75371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520500
GitOrigin-RevId: 27f132fcd87230955f6261dd44a26072c381baba
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/535342
Original-Revision: 5ccd222a7be2255d5d54c5ec1f4e83ecdc4f897e
Change-Id: I5fbad2398bfd7334570989854c4859bfd9586b84