[roll] Roll fuchsia [kernel][x86] Remove `amd_retpoline` global
This value gave whether we were using the AMD-appropriate retpoline
patch alternative (printed in `x86_feature_debug()`). The code-patching
infrastructure now prints this information out on boot instead and
nothing sets this anymore. Given that we do log the relevant information
(and that it would be a bit awkward to go back to setting this variable
in feature.cc), we remove it in this change (as was originally meant to
be done).

Original-Bug: 77306
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535381
Original-Revision: fa0c3e7b4b5c5cc77f0f959aef95ef2baa9b3a56
GitOrigin-RevId: 7ca5657404f95eedd07102fb525077c2377bf870
Change-Id: I4b6d0e923b8ea7368c14c4e94f3ac74b37303248