[roll] Roll infra/recipes [build] Check fint results vs. affected_tests.py
fxrev.dev/520500 rewrites the affected_tests.py logic in fint. To give
us confidence that this will not regress affected tests analysis when
migrating away from the battle-tested affected_tests.py script, compare
the results of the two tools and emit a dummy step if the results
differ, to make it possible to run Buildbucket BigQuery queries to
determine whether there are instances in which the two tools disagree.

I'll wait until this has run for a week or so with no disagreements
between the two tools before deleting affected_tests.py and making
fint's affected tests analysis be the source of truth.

I also did a small refactoring to pass around a list of affected test
names instead of a path to a file containing the affected test names, to
avoid reading the file multiple times and to consolidate assumptions
about the file's schema.

Original-Bug: 75371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/531422
Original-Revision: 8d34bfe07a4c15a61085f361ce40660a40a3f6ee
GitOrigin-RevId: 9a11a2cde990a64764b618d968e3c2e6bbeddd88
Change-Id: I833ea0f4e600cbec589ed8b7d7a6a4df1d4b81ca