[roll] Roll fuchsia [kernel_cmdline] Disable virtio_ethernet in GCE
GCE VMs are experiencing kernel lockups due to a bad interaction between
the virtio_ethernet driver and Vanadium, see the attached bug for more
context. Note that this will break the core.x64-gce builder, but we
cannot pass this in from the builder definition because we need the
kernel arg in the image.

Original-Bug: b/188675754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534682
Original-Revision: 401311d6496c661867ff301d18c54fb3e0278633
GitOrigin-RevId: 2a4e3c96da235e78baec38a94d86502b0e6d0e1f
Change-Id: I8dd28a68907b8b2fedea96b993e175826e7cec32