[roll] Roll fuchsia [ftl] Reject map pages with all 0xff in the second half of OOB.
This is a reland that adds the changes to the ftl-mtd to override the
NdmBaseDriver methods for supplying page and spare sizes. Original
change was fxr/529803

This rejects map pages with all 0xff in the second half of the OOB in
order to reject partial page writes which result in wear count listed as
the max value and the validity bits are 0xff, though validity bits are
currently all 0xff.

This also starts writing a form of checksum to the last validity byte to
ensure that the write completed, and to spot any unexpected 0xff
patterns in the oob that may have resulted from an incomplete write.

Original-Bug: b/179400686
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534962
Original-Revision: b1743da07826429c4a8495e4f8cad174903941b9
GitOrigin-RevId: 64aab25dc4c87a13a614cc0431e061ab963e3613
Change-Id: I14c8d87ade787ad6b6cbef9e05b1c29b822b27e6