[roll] Roll fuchsia [display] Add flag to disable kernel framebuffer.
Forcing all virtcon images to be set as kernel framebuffer
limits that formats that can be used and significantly
impacts performance when virtcon is not in single framebuffer
mode.

This adds a flag that disable kernel framebuffer support
for configurations where that is desirable.

Original-Bug: 76124
Test: display-test (with driver.display.disable-kernel-framebuffer=1)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527241
Original-Revision: 637b3cb8db137e419d9e0af612692fae9ac2cebc
GitOrigin-RevId: 277f73fae5625e448fc30d10b0a135056bd8fd62
Change-Id: I06b5c9072885ba6f19e4c35728c61720553bd6a1