[roll] Roll fuchsia [fidl][dart] Receive-side handle rights checking
This CL adds receive-side handle rights checking in Dart. This means
that handle rights and type are checked on the receiving side of the
channel.

The tests in this CL just test end-to-end behavior and do no isolate
checks that the behavior is happening on the receiving side (that will
happen later). In addition, they only check that the _checkHandleRights
method is called by testing that wrong handle types are caught because
detailed tests of type and rights variations will be covered by GIDL
decode tests.

Some upcoming changes:
- zx_handle_replace is not currently supported in Dart, but it will be
supported after the next Flutter roll. Because of this, rights reduction
cannot be supported at this time.
- GIDL support will be used to validate the decoding logic. However,
this involves significant changes and will be done in a separate CL.
- Send-side handle rights checking needs to be completed as well.

Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534541
Original-Revision: df96c1c615516e7b53269abce3003af6dbb9222e
GitOrigin-RevId: 3ac400e32b84b9a04d0ee3f5657e1de2dec7d92f
Change-Id: Ib481307f3f72ac6ed5237cf20536268b96f546ff