[roll] Roll fuchsia [ffx] manual targets never enter disconnected
A manual target is non-discoverable. The disconnected state is
essentially a mis-nomer that actually means "needs to be rediscovered",
as actions such as connecting to the target, or listing in target
listings are dependent on the state being not-disconnected.

update_connection_state enforces this transition.

Original-Bug: 76632
Original-Fixed: 76426
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534363
Original-Revision: 0ab08d20564e16568d980af9e7091ccafeda054f
GitOrigin-RevId: 5e260412ba6a6635ea38fbb8253bbf4a3e7ce5db
Change-Id: If4fc8829156f733479fcc169276da9b727239ade