[roll] Roll infra/recipes [fint] Add build artifacts fields for affected tests
We're moving the affected_tests.py logic into fint and need a way to
expose the information from fint back to recipes.

Originally this was included along with the fint logic changes in
fxrev.dev/520500 but that change needed to be reverted, which ended up
being very messy because subsequent recipe changes that depended on
these new fields then had to be reverted.

Landing the proto change separately in advance will make it easier to
revert the logic change again, if necessary.

Original-Bug: 75371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536043
GitOrigin-RevId: 0a063e45b1a382d197c097d5806adf700459c03a
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/536002
Original-Revision: 2ecf0809590647c7c3904d2d8364fd0a08e4f280
Change-Id: Ie7a85b87e578dce75a03da0901d90aad606bbd29