[roll] Roll infra/recipes [test_owners] Upload to BigQuery
Having test owners data easily accessible in BigQuery will be useful for
CTS testing and other efforts.

For now we just hardcode the table name in the recipe, since we already
do that for the ResultDB and CQ tables in the same recipe, and the
recipe is kind of a one-off so it doesn't matter too much if some things
are hard-coded. If it becomes an issue we can instead pass the table
name and GCP project name as recipe properties.

Original-Bug: 77227
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/535364
Original-Revision: 8deb66ac9ec10042fc6c48b09d097d82284530b3
GitOrigin-RevId: d9f11b7bc4c6e665042818fe94c7b0b548257e40
Change-Id: I90cb44f7aea702d935a8f7b0cee177d0d78b53dd