[roll] Roll fuchsia [fidl][summary] Change API signatures
Reviewers note: the only real change is in `fidl_library.gni`.  The
`.api` file changes are all changes of the library hash value.

The API signatures are now computed based on the API summary files, not
the normalized form of the library.  This should be a more stable format
for API summaries, changing the API signature only when real API
impacting changes happen.

As result of the change of the signature method, all API signatures are
changing.  This is inevitable, and may need to be undone in case this
change backfires somehow.

We'll start checking in the API summary files alongside the respective
APIs, but for API summary checks this is not yet needed.  No parts of
prior fixtures are removed yet, everything is flag-guarded.

Original-Bug: 68952
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533141
Original-Revision: 026ede2f76e41faf3fc94d4df7fdf2122a940bca
GitOrigin-RevId: 2c3daf0a2c55b214a7e0518ad8be43ee60488b80
Change-Id: I4d9e7a0c522185cba384d0de34bbf226e837e1a5